### 2024-09-21 #7260 @editorialbot
[PRE REVIEW]: DiscreteEntropy.jl: Entropy Estimation of Discrete Random Variables with Julia
**Submitting author:** @kellino (<a href="http://orcid.org/0000-0002-5368-6769">David Kelly</a>)
**Repository:** https://github.com/kellino/DiscreteEntropy.jl
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.2.0
