+++ paper/paper.bib
@@ -40,6 +40,7 @@ eprint = {https://doi.org/10.1137/1104033},
   publisher = {Copyright Cambridge University Press},
   timestamp = {2007-05-24T14:43:04.000+0200},
   title = {Information Theory, Inference, and Learning Algorithms},
+  doi = {10.1109/tit.2004.834752},
   year = 2003
 }
 
@@ -51,6 +52,7 @@ eprint = {https://doi.org/10.1137/1104033},
   number={5},
   pages={561},
   year={2021},
+  doi={10.3390/e23050561},
   publisher={MDPI}
 }
 
@@ -74,13 +76,19 @@ journal = {Environmental and Ecological Statistics},
 doi = {10.1023/A:1026096204727}
 }
 
-@misc{nemenman2002entropy,
-      title={Entropy and inference, revisited},
-      author={Ilya Nemenman and Fariel Shafee and William Bialek},
-      year={2002},
-      eprint={physics/0108025},
-      archivePrefix={arXiv},
-      primaryClass={physics.data-an}
+
+@inproceedings{10.5555/2980539.2980601,
+  author = {Nemenman, Ilya and Shafee, Fariel and Bialek, William},
+  title = {Entropy and inference, revisited},
+  year = {2001},
+  publisher = {MIT Press},
+  address = {Cambridge, MA, USA},
+  booktitle = {Proceedings of the 14th International Conference on Neural Information Processing Systems: Natural and Synthetic},
+  pages = {471–478},
+  numpages = {8},
+  location = {Vancouver, British Columbia, Canada},
+  series = {NIPS'01},
+  doi={10.7551/mitpress/1120.003.0065},
 }
 
 @article{zhang,
