### 2024-10-01 #7260 @Nikoleta-v3
Hello! Thank you for the invite. FYI @danielskatz @kellino, I am not an expert in Julia, I have barely used it, but I’m interested in learning more about the language, especially when it comes to packaging. I’d be very happy to edit the submission.

I’ll make sure to find reviewers who are experts in Julia, haha 😄
