### 2024-10-01 #7260 @kellino
@Nikoleta-v3 I've got to confess that I'm no expert on Julia either (this is actually my first package in the language). Hopefully it should be fairly easy to install and use. I'm open to all suggestions for improvements
