### 2024-10-01 #7260 @Nikoleta-v3
Well then, it will be a learning experience for both of us! Before we begin a few things @kellino.

(-) After a quick look at the repository, one file that seems to be missing is the contribution guidelines. Here’s an example of a contributions file from a recent publication I was editing: https://forgemia.inra.fr/pherosensor/pherosensor-toolbox/-/blob/main/CONTRIBUTING.rst?ref_type=heads. It's clear and to the point, so you might want to use this as a guide. There are also more detailed examples, such as: https://github.com/damar-wicaksono/uqtestfuns/blob/dev/CONTRIBUTING.MD

Or you could keep it simple, like this one-sentence example: https://github.com/RECeSS-EU-Project/stanscofi/

I’ll leave it up to you. Could you please make sure to include a section on how to run the tests?

(-) With a quick read through the paper, I noticed a few typos. Could you please correct them?

- `is a Julia package` (Line 6)
- `It is well known` (Line 11)
- `This difficulty has led` (Line 12)
- `for those who wish` (Line 37)

(-) Once you’ve addressed the above comments, could you please visit: https://reviewers.joss.theoj.org and recommend a few reviewers?

Thank you!
