### 2024-10-02 #7260 @kellino
@Nikoleta-v3 I've fixed the typos in the paper (thank you for spotting these) and added a CONTRIBUTING.md, AUTHORS.md and CHANGELOG.md. All of these in paper branch for the moment. I've borrowed heavily from [pherosensor](https://forgemia.inra.fr/pherosensor/pherosensor-toolbox/-/blob/main/CONTRIBUTING.rst?ref_type=heads) in terms of stucture. Should I give them a quick acknowledgement at the end? I'll have another look for reviewers either today or tomorrow, but having flicked through the list before, nobody was an obvious fit.
