### 2024-10-02 #7260 @Nikoleta-v3
Thank you @kellino!

> I've borrowed heavily from [pherosensor](https://forgemia.inra.fr/pherosensor/pherosensor-toolbox/-/blob/main/CONTRIBUTING.rst?ref_type=heads) in terms of stucture. Should I give them a quick acknowledgement at the end?

No harm in acknowledging them 😄 but it's really up to you!

Some of the authors here: https://github.com/openjournals/joss-reviews/issues/7260#issuecomment-2367632334 could also be potential reviewers 😄
