### 2024-10-03 #7260 @kellino
@Nikoleta-v3  the projects with the most obvious overlap are the first two in https://github.com/openjournals/joss-reviews/issues/7260#issuecomment-2367632334, though neither of those projects are in Julia. If the author of https://github.com/madvn/infotheory (madvn) agrees to review, perhaps someone else familar with Julia can also be brought on board?

Failing that, any of the  reviewers for the first two papers (ajgates42, artemyk, anotherjoshsmith) should definitely have the expertise from the information theory point of view. `DiscreteEntropy.jl` is intended to be useful rather than groundbreaking. So maybe a reviewer from a Julia project (eg ziyiyin97, dannys4) might also consider lending their expertise.
