### 2024-10-11 kellino/DiscreteEntropy.jl#11 @nluetts
Testing against Contreras Rodríguez et al. 2021
Hi @kellino :wave:

this is Nils, I am opening this in response to the review over at JOSS
(https://github.com/openjournals/joss-reviews/issues/7334).

From your JOSS paper draft I got the impression that the paper from Contreras
Rodrı́guez et al. (https://doi.org/10.3390/e23050561) is quite central for your
package, so I thought testing your package against the estimated entropies
published there would be a good thing to do (which does not yet seem to be
covered by your test cases, although they are quite comprehensive otherwise).

In the paper, they estimate entropies for byte sequences they generated on
a Linux machine with `/dev/urandom`, so I did the same and tried to
reproduce their Table 1 with `DiscreteEntropy.jl`.

I used this Julia environment:

```julia
[nluetts@fedora DiscreteEntropy-JOSS-Review]$ julia --project
               _
   _       _ _(_)_     |  Documentation: https://docs.julialang.org
  (_)     | (_) (_)    |
   _ _   _| |_  __ _   |  Type "?" for help, "]?" for Pkg help.
  | | | | | | |/ _` |  |
  | | |_| | | | (_| |  |  Version 1.10.5 (2024-08-27)
 _/ |\__'_|_|_|\__'_|  |  Official https://julialang.org/ release
|__/                   |

(DiscreteEntropy-JOSS-Review) pkg> status
Status `/var/home/nluetts/Repos/DiscreteEntropy-JOSS-Review/Project.toml`
  [940648b6] DiscreteEntropy v0.2.0
  [08abe8d2] PrettyTables v2.4.0
  [10745b16] Statistics v1.10.0
```

And the following script to do the comparison:

```julia
using DiscreteEntropy
using Statistics
using PrettyTables

function main()
    results = Dict{String,Vector{Float64}}()
    for est_label in keys(REF_VALUES)
        println("###############################################################################")
        println()
        est = ESTIMATORS[est_label]
        if isnothing(est)
            @warn "No estimator for \"$(est_label)\", skipping ..."
            # so that results table is not missing a row from the paper
            results[est_label] = repeat([NaN64], 12)
            continue
        end

        # less draws for slower estimators
        draws = est == BUB ? 100 : 1000
        draws = est == NSB ? 25 : draws

        println("Running comparison for $est ...")

        mean_entropies = Float64[]
        for exp in 3:14
            println("Sample size: $(2^exp)")
            try
                entropy = compare_contreras_rodriguez(est, exp; draws=draws)
                push!(mean_entropies, entropy)
                # index in list of entropy values = exp - 2
                # because Contreras Rodríguez et al. starts with
                # sample size 2^3 = 8
                ref_entropy = REF_VALUES[est_label][exp-2]
                println("Reference mean entropy: $ref_entropy, DiscreteEntropy.jl: $entropy")
            catch e
                println("ERROR running estimation: $e")
                push!(mean_entropies, NaN64)
            end
        end
        results[est_label] = mean_entropies
    end

    labels = SORTED_ESTIMATOR_LABELS
    n = length(labels)
    # since PrettyTables cannot transpose, we have
    # to do some work ourselves here ...
    estimates = [round(results[k][exp-2], digits=3) for exp in 3:14 for k in labels]
    deviations = [
        round(results[k][exp-2] - REF_VALUES[k][exp-2], digits=2)
        for exp in 3:14 for k in labels
    ]
    raw_table_results = hcat(labels, reshape(estimates, n, :))
    raw_table_deviations = hcat(labels, reshape(deviations, n, :))
    header = vcat(["Estimator"], [2^n for n in 3:14])

    # print out tables
    pretty_table(raw_table_results; header=header),
    pretty_table(raw_table_deviations; header=header)
end

# From Contreras Rodríguez et al.,
# Entropy 2021, 23, 561. https://doi.org/10.3390/e23050561
# Table 2: "Estimated mean of estimates for byte samples."
REF_VALUES = Dict(
    "CS" => [6.142, 7.635, 8.251, 8.126, 8.064, 8.099, 8.116, 8.003, 7.914, 7.940, 7.963, 7.974],
    "ML" => [2.969, 3.943, 4.880, 5.768, 6.550, 7.178, 7.591, 7.809, 7.908, 7.955, 7.977, 7.989],
    "MM" => [3.590, 4.598, 5.536, 6.397, 7.113, 7.633, 7.902, 7.985, 7.998, 7.999, 8.000, 8.000],
    "SG" => [3.999, 4.470, 5.136, 5.884, 6.598, 7.194, 7.595, 7.809, 7.908, 7.955, 7.977, 7.989],
    "SHR" => [7.940, 7.971, 7.982, 7.991, 7.995, 7.998, 7.999, 7.999, 8.000, 8.000, 8.000, 8.000],
    "JEF" => [7.947, 7.902, 7.834, 7.751, 7.690, 7.696, 7.767, 7.854, 7.919, 7.957, 7.978, 7.989],
    "LAP" => [7.983, 7.969, 7.943, 7.904, 7.859, 7.832, 7.842, 7.883, 7.928, 7.960, 7.979, 7.989],
    "MIN" => [5.048, 5.385, 5.835, 6.357, 6.875, 7.326, 7.643, 7.823, 7.912, 7.956, 7.978, 7.989],
    "BUB" => [7.008, 7.021, 7.476, 7.732, 7.868, 8.061, 7.896, 7.971, 7.983, 7.985, 7.985, 7.985],
    "NSB" => [6.427, 7.065, 7.474, 7.726, 7.855, 7.922, 7.953, 7.969, 7.977, 7.981, 7.983, 7.984],
    "Unseen" => [5.202, 6.751, 7.739, 7.925, 7.945, 7.957, 7.950, 7.948, 7.953, 7.960, 7.966, 7.971],
    "Zhang" => [3.690, 4.696, 5.627, 6.478, 7.178, 7.674, 7.916, 7.979, 7.985, 7.985, 7.985, 7.985],
    "CJ" => [5.349, 7.028, 8.089, 8.093, 8.004, 7.996, 7.987, 7.985, 7.985, 7.985, 7.985, 7.985],
    "UnveilJ" => [3.036, 4.124, 5.316, 6.645, 7.749, 7.95, 7.855, 7.832, 7.894, 7.940, 7.963, 7.974],
    "BN" => [2.963, 4.433, 5.815, 6.981, 7.726, 7.889, 7.41, 6.662, 6.115, 5.829, 5.686, 5.616],
    "GSB" => [4.646, 5.613, 6.464, 7.182, 7.670, 7.921, 7.981, 7.985, 7.985, 7.985, 7.985, 7.985],
    "SHU" => [4.347, 5.329, 6.210, 6.978, 7.541, 7.868, 7.972, 7.985, 7.985, 7.985, 7.985, 7.985],
    "CDM" => [6.165, 6.908, 7.391, 7.689, 7.844, 7.922, 7.961, 7.981, 7.990, 7.995, 7.998, 7.999],
)

# same sorting as in paper (not preserved in Dict `REF_VALUES`)
SORTED_ESTIMATOR_LABELS = ["CS", "ML", "MM", "SG", "SHR", "JEF", "LAP", "MIN", "BUB", "NSB", "Unseen", "Zhang", "CJ", "UnveilJ", "BN", "GSB", "SHU", "CDM"]

ESTIMATORS = Dict(
    "CS" => ChaoShen,
    "ML" => MaximumLikelihood,
    "MM" => MillerMadow,
    "SG" => SchurmannGrassberger,
    "SHR" => Shrink,
    "JEF" => Jeffrey,
    "LAP" => LaPlace,
    "MIN" => Minimax,
    "BUB" => BUB,
    "NSB" => NSB,
    "Unseen" => Unseen,
    "Zhang" => Zhang,
    # Contreras Rodríguez et al. seem to have a typo here, CJ is not mentioned
    # in Table 1, only CWJ, so I am not sure this mapping is correct.
    "CJ" => ChaoWangJost,
    "UnveilJ" => JackknifeMLE,
    "BN" => Bonachela,
    "GSB" => Grassberger,
    "SHU" => Schurmann,
    "CDM" => nothing # Cannot figure out which one this is
)

function read_dev_urandom_bytes(n::UInt)
    bytes = open("/dev/urandom", "r") do io
        read(io, n)
    end
    bytes
end

read_dev_urandom_bytes(n) = read_dev_urandom_bytes(convert(UInt, n))

function compare_contreras_rodriguez(::Type{T}, bytes_exponent::UInt8; draws=1000) where {T<:AbstractEstimator}
    @assert 3 <= bytes_exponent <= 14 "No reference values available for sample size $(2^bytes_exponent)"
    sample_size = 2^bytes_exponent
    mean(
        estimate_h(read_dev_urandom_bytes(sample_size) |> svector |> from_samples, T)
        for _ in 1:draws
    ) |> to_bits
end

compare_contreras_rodriguez(::Type{T}, bytes_exponent; draws=1000) where {T<:AbstractEstimator} =
    compare_contreras_rodriguez(T, convert(UInt8, bytes_exponent), draws=draws)
```

The script generates the following output (scroll to the end to find the two relevant tables):

```
[nluetts@fedora DiscreteEntropy-JOSS-Review]$ julia --project --load comparison.jl -e "main()"
###############################################################################

Running comparison for BUB ...
Sample size: 8
Reference mean entropy: 7.008, DiscreteEntropy.jl: 3.614995850373481
Sample size: 16
Reference mean entropy: 7.021, DiscreteEntropy.jl: 4.7899950608286606
Sample size: 32
Reference mean entropy: 7.476, DiscreteEntropy.jl: 6.796319151130138
Sample size: 64
Reference mean entropy: 7.732, DiscreteEntropy.jl: 7.572799602723234
Sample size: 128
Reference mean entropy: 7.868, DiscreteEntropy.jl: 7.9792128580842085
Sample size: 256
Reference mean entropy: 8.061, DiscreteEntropy.jl: 8.058072331938295
Sample size: 512
Reference mean entropy: 7.896, DiscreteEntropy.jl: 8.001327375927534
Sample size: 1024
Reference mean entropy: 7.971, DiscreteEntropy.jl: 7.99473597067706
Sample size: 2048
Reference mean entropy: 7.983, DiscreteEntropy.jl: 8.000663690919412
Sample size: 4096
Reference mean entropy: 7.985, DiscreteEntropy.jl: 7.9938103137202425
Sample size: 8192
Reference mean entropy: 7.985, DiscreteEntropy.jl: 7.999977554911958
Sample size: 16384
Reference mean entropy: 7.985, DiscreteEntropy.jl: 8.000005950701526
###############################################################################

Running comparison for ChaoWangJost ...
Sample size: 8
Reference mean entropy: 5.349, DiscreteEntropy.jl: 5.360278389970844
Sample size: 16
Reference mean entropy: 7.028, DiscreteEntropy.jl: 7.041184521071105
Sample size: 32
Reference mean entropy: 8.089, DiscreteEntropy.jl: 8.118867698781537
Sample size: 64
Reference mean entropy: 8.093, DiscreteEntropy.jl: 8.08072337036545
Sample size: 128
Reference mean entropy: 8.004, DiscreteEntropy.jl: 8.01966232891067
Sample size: 256
Reference mean entropy: 7.996, DiscreteEntropy.jl: 8.008443797783935
Sample size: 512
Reference mean entropy: 7.987, DiscreteEntropy.jl: 8.000142338203254
Sample size: 1024
Reference mean entropy: 7.985, DiscreteEntropy.jl: 8.000489112501155
Sample size: 2048
Reference mean entropy: 7.985, DiscreteEntropy.jl: 8.00003292715623
Sample size: 4096
Reference mean entropy: 7.985, DiscreteEntropy.jl: 7.999791811121944
Sample size: 8192
Reference mean entropy: 7.985, DiscreteEntropy.jl: 7.999968198405732
Sample size: 16384
Reference mean entropy: 7.985, DiscreteEntropy.jl: 7.9999736567289865
###############################################################################

Running comparison for Minimax ...
Sample size: 8
Reference mean entropy: 5.048, DiscreteEntropy.jl: 2.9761034489686033
Sample size: 16
Reference mean entropy: 5.385, DiscreteEntropy.jl: 3.9442625156600024
Sample size: 32
Reference mean entropy: 5.835, DiscreteEntropy.jl: 4.89343609826837
Sample size: 64
Reference mean entropy: 6.357, DiscreteEntropy.jl: 5.783872751201355
Sample size: 128
Reference mean entropy: 6.875, DiscreteEntropy.jl: 6.565523196303405
Sample size: 256
Reference mean entropy: 7.326, DiscreteEntropy.jl: 7.196160029923576
Sample size: 512
Reference mean entropy: 7.643, DiscreteEntropy.jl: 7.606140946314727
Sample size: 1024
Reference mean entropy: 7.823, DiscreteEntropy.jl: 7.819207596467573
Sample size: 2048
Reference mean entropy: 7.912, DiscreteEntropy.jl: 7.912123086208508
Sample size: 4096
Reference mean entropy: 7.956, DiscreteEntropy.jl: 7.956119491898474
Sample size: 8192
Reference mean entropy: 7.978, DiscreteEntropy.jl: 7.978118017008263
Sample size: 16384
Reference mean entropy: 7.989, DiscreteEntropy.jl: 7.988922826292928
###############################################################################

Running comparison for Grassberger ...
Sample size: 8
Reference mean entropy: 4.646, DiscreteEntropy.jl: 4.475872575973066
Sample size: 16
Reference mean entropy: 5.613, DiscreteEntropy.jl: 5.395986343698666
Sample size: 32
Reference mean entropy: 6.464, DiscreteEntropy.jl: 6.260004822058096
Sample size: 64
Reference mean entropy: 7.182, DiscreteEntropy.jl: 6.980796744618755
Sample size: 128
Reference mean entropy: 7.67, DiscreteEntropy.jl: 7.5182629313850935
Sample size: 256
Reference mean entropy: 7.921, DiscreteEntropy.jl: 7.827305281488433
Sample size: 512
Reference mean entropy: 7.981, DiscreteEntropy.jl: 7.961604461074284
Sample size: 1024
Reference mean entropy: 7.985, DiscreteEntropy.jl: 7.995547205459536
Sample size: 2048
Reference mean entropy: 7.985, DiscreteEntropy.jl: 8.000232539142349
Sample size: 4096
Reference mean entropy: 7.985, DiscreteEntropy.jl: 8.000359491911881
Sample size: 8192
Reference mean entropy: 7.985, DiscreteEntropy.jl: 7.999967625418502
Sample size: 16384
Reference mean entropy: 7.985, DiscreteEntropy.jl: 7.9998609625723525
###############################################################################

Running comparison for Jeffrey ...
Sample size: 8
Reference mean entropy: 7.947, DiscreteEntropy.jl: 2.972254402351613
Sample size: 16
Reference mean entropy: 7.902, DiscreteEntropy.jl: 3.9547463784273966
Sample size: 32
Reference mean entropy: 7.834, DiscreteEntropy.jl: 4.8988101907857375
Sample size: 64
Reference mean entropy: 7.751, DiscreteEntropy.jl: 5.796691217187817
Sample size: 128
Reference mean entropy: 7.69, DiscreteEntropy.jl: 6.595211026669902
Sample size: 256
Reference mean entropy: 7.696, DiscreteEntropy.jl: 7.2428089132809506
Sample size: 512
Reference mean entropy: 7.767, DiscreteEntropy.jl: 7.6545957368144855
Sample size: 1024
Reference mean entropy: 7.854, DiscreteEntropy.jl: 7.844389558269971
Sample size: 2048
Reference mean entropy: 7.919, DiscreteEntropy.jl: 7.918733289952589
Sample size: 4096
Reference mean entropy: 7.957, DiscreteEntropy.jl: 7.957507646842208
Sample size: 8192
Reference mean entropy: 7.978, DiscreteEntropy.jl: 7.978231591590087
Sample size: 16384
Reference mean entropy: 7.989, DiscreteEntropy.jl: 7.988930604764357
###############################################################################

Running comparison for LaPlace ...
Sample size: 8
Reference mean entropy: 7.983, DiscreteEntropy.jl: 2.9736909829661333
Sample size: 16
Reference mean entropy: 7.969, DiscreteEntropy.jl: 3.9533018294946505
Sample size: 32
Reference mean entropy: 7.943, DiscreteEntropy.jl: 4.903365419183482
Sample size: 64
Reference mean entropy: 7.904, DiscreteEntropy.jl: 5.808311331087142
Sample size: 128
Reference mean entropy: 7.859, DiscreteEntropy.jl: 6.621930848735704
Sample size: 256
Reference mean entropy: 7.832, DiscreteEntropy.jl: 7.27588208048709
Sample size: 512
Reference mean entropy: 7.842, DiscreteEntropy.jl: 7.692674035956105
Sample size: 1024
Reference mean entropy: 7.883, DiscreteEntropy.jl: 7.869540223819323
Sample size: 2048
Reference mean entropy: 7.928, DiscreteEntropy.jl: 7.927973374245435
Sample size: 4096
Reference mean entropy: 7.96, DiscreteEntropy.jl: 7.959961271968038
Sample size: 8192
Reference mean entropy: 7.979, DiscreteEntropy.jl: 7.9786682763551084
Sample size: 16384
Reference mean entropy: 7.989, DiscreteEntropy.jl: 7.989102788616696
###############################################################################

Running comparison for Shrink ...
Sample size: 8
Reference mean entropy: 7.94, DiscreteEntropy.jl: 2.9808091479101586
Sample size: 16
Reference mean entropy: 7.971, DiscreteEntropy.jl: 3.953894179273727
Sample size: 32
Reference mean entropy: 7.982, DiscreteEntropy.jl: 4.911835316261442
Sample size: 64
Reference mean entropy: 7.991, DiscreteEntropy.jl: 5.821070473660774
Sample size: 128
Reference mean entropy: 7.995, DiscreteEntropy.jl: 6.655003767983381
Sample size: 256
Reference mean entropy: 7.998, DiscreteEntropy.jl: 7.329850657777553
Sample size: 512
Reference mean entropy: 7.999, DiscreteEntropy.jl: 7.758420052803192
Sample size: 1024
Reference mean entropy: 7.999, DiscreteEntropy.jl: 7.929509738707598
Sample size: 2048
Reference mean entropy: 8.0, DiscreteEntropy.jl: 7.966671297879019
Sample size: 4096
Reference mean entropy: 8.0, DiscreteEntropy.jl: 7.983602410097988
Sample size: 8192
Reference mean entropy: 8.0, DiscreteEntropy.jl: 7.992357342562944
Sample size: 16384
Reference mean entropy: 8.0, DiscreteEntropy.jl: 7.9966348604391735
###############################################################################

Running comparison for SchurmannGrassberger ...
Sample size: 8
Reference mean entropy: 3.999, DiscreteEntropy.jl: 2.971736967454633
Sample size: 16
Reference mean entropy: 4.47, DiscreteEntropy.jl: 3.944955929980586
Sample size: 32
Reference mean entropy: 5.136, DiscreteEntropy.jl: 4.8825236480803875
Sample size: 64
Reference mean entropy: 5.884, DiscreteEntropy.jl: 5.7679317080531565
Sample size: 128
Reference mean entropy: 6.598, DiscreteEntropy.jl: 6.549177687335087
Sample size: 256
Reference mean entropy: 7.194, DiscreteEntropy.jl: 7.176950412445397
Sample size: 512
Reference mean entropy: 7.595, DiscreteEntropy.jl: 7.591266466353353
Sample size: 1024
Reference mean entropy: 7.809, DiscreteEntropy.jl: 7.808764456449525
Sample size: 2048
Reference mean entropy: 7.908, DiscreteEntropy.jl: 7.908099032251887
Sample size: 4096
Reference mean entropy: 7.955, DiscreteEntropy.jl: 7.954910030058388
Sample size: 8192
Reference mean entropy: 7.977, DiscreteEntropy.jl: 7.977386074854952
Sample size: 16384
Reference mean entropy: 7.989, DiscreteEntropy.jl: 7.988781719453369
###############################################################################

Running comparison for Zhang ...
Sample size: 8
Reference mean entropy: 3.69, DiscreteEntropy.jl: 3.70057718590882
Sample size: 16
Reference mean entropy: 4.696, DiscreteEntropy.jl: 4.708114791136625
Sample size: 32
Reference mean entropy: 5.627, DiscreteEntropy.jl: 5.640825481842562
Sample size: 64
Reference mean entropy: 6.478, DiscreteEntropy.jl: 6.489832547884759
Sample size: 128
Reference mean entropy: 7.178, DiscreteEntropy.jl: 7.185693307022013
Sample size: 256
Reference mean entropy: 7.674, DiscreteEntropy.jl: 7.685618818498393
Sample size: 512
Reference mean entropy: 7.916, DiscreteEntropy.jl: 7.931552076659486
Sample size: 1024
Reference mean entropy: 7.979, DiscreteEntropy.jl: 7.995478619546022
Sample size: 2048
Reference mean entropy: 7.985, DiscreteEntropy.jl: 8.00011685326433
Sample size: 4096
Reference mean entropy: 7.985, DiscreteEntropy.jl: 8.0002027748647
Sample size: 8192
Reference mean entropy: 7.985, DiscreteEntropy.jl: 7.999933474194657
Sample size: 16384
Reference mean entropy: 7.985, DiscreteEntropy.jl: 7.999973102305582
###############################################################################

Running comparison for Bonachela ...
Sample size: 8
Reference mean entropy: 2.963, DiscreteEntropy.jl: 3.258906518604906
Sample size: 16
Reference mean entropy: 4.433, DiscreteEntropy.jl: 5.008529414146986
Sample size: 32
Reference mean entropy: 5.815, DiscreteEntropy.jl: 6.828142608055729
Sample size: 64
Reference mean entropy: 6.981, DiscreteEntropy.jl: 8.479980699266957
Sample size: 128
Reference mean entropy: 7.726, DiscreteEntropy.jl: 9.7162166211824
Sample size: 256
Reference mean entropy: 7.889, DiscreteEntropy.jl: 10.242244993398158
Sample size: 512
Reference mean entropy: 7.41, DiscreteEntropy.jl: 9.956818207508285
Sample size: 1024
Reference mean entropy: 6.662, DiscreteEntropy.jl: 9.221113291455794
Sample size: 2048
Reference mean entropy: 6.115, DiscreteEntropy.jl: 8.632933243927424
Sample size: 4096
Reference mean entropy: 5.829, DiscreteEntropy.jl: 8.316754014736388
Sample size: 8192
Reference mean entropy: 5.686, DiscreteEntropy.jl: 8.158346741992037
Sample size: 16384
Reference mean entropy: 5.616, DiscreteEntropy.jl: 8.079233029603056
###############################################################################

Running comparison for MillerMadow ...
Sample size: 8
Reference mean entropy: 3.59, DiscreteEntropy.jl: 3.592985854164966
Sample size: 16
Reference mean entropy: 4.598, DiscreteEntropy.jl: 4.5973992117016325
Sample size: 32
Reference mean entropy: 5.536, DiscreteEntropy.jl: 5.5378266814833585
Sample size: 64
Reference mean entropy: 6.397, DiscreteEntropy.jl: 6.3944223125125115
Sample size: 128
Reference mean entropy: 7.113, DiscreteEntropy.jl: 7.116156496268463
Sample size: 256
Reference mean entropy: 7.633, DiscreteEntropy.jl: 7.6320174448968015
Sample size: 512
Reference mean entropy: 7.902, DiscreteEntropy.jl: 7.903183965765673
Sample size: 1024
Reference mean entropy: 7.985, DiscreteEntropy.jl: 7.984768837092095
Sample size: 2048
Reference mean entropy: 7.998, DiscreteEntropy.jl: 7.99727572744489
Sample size: 4096
Reference mean entropy: 7.999, DiscreteEntropy.jl: 7.999653223504991
Sample size: 8192
[…]
