### 2024-10-11 kellino/DiscreteEntropy.jl#11 @kellino
Thanks @nluetts. I'll do my best to answer these questions, but some will need more investigation. First, I have not implemented "CDM". It doesn't seem much used in the literature and is largely (as far as I can tell) superceded in the literature by PYM. I'll probably add it to the library at a later date if there is any demand of it.

I can't comment on the quality of the work in Rodriguez et al. and I have not tried to match against their results. I really only used it as a (partial) basis for what to implement. As far as possible, I went back to the authors original implementations (where they existed) and compared output against them.

NSB is an interesting case, as the author's implementation gives different results from the python implementation used by Rodriguez et al., which gives different results from `DiscreteEntropy.jl`. I cannot claim with certainty that my version is correct, but it does most faithfully copy the maths as presented in the paper. NSB (and some of the other bayesian estimators) uses integration: this could be a major source of divergence. To avoid complex rewriting to avoid overflow (as found in other NSB implementations) I used Julia's arbitrary precision floats. Integrating over these could easily introduce some of the deviation that you see. The question then becomes was the lack of precision considered by the authors when presenting their maths. My guess is, it was not, and therefore deviations are due to floating point imperfections. I noticed while writing the code that these do add up very quickly.

Also, at small sample sizes, entropy estimators vary tremendously from run to run. 1000 iterations may not be enough for stability at very small sizes. I note that as sample size increases, the tables converge. BN is interesting though. It is specifically for small datasets (I will add this information into the docs): divergence increases as more samples are added, which is the behaviour I think I would expect to see and at small sizes we again have the stability problem. Looking at the code in `frequentist.jl` I can't immediately see an error compared against the formula in the docs.

I will add a check for sample size on `NSB` < 64, though that sounds more like a bug. I'm not sure what's happening with the Unseen, please give me a few days to figure it out.
