### 2024-10-11 kellino/DiscreteEntropy.jl#11 @kellino
I've found the error in NSB: whenever the input has no coincidences (ie everything is seen only once), it fails. I need to figure out why that is happening and how to fix it. Thanks for the spot!
