### 2024-10-11 #7334 @niyiyu
@kellino Thank you for submitting this Julia package. I only have several minor suggestions to meet the JOSS criteria.

1. `Equation (1)` is not correctly rendered in the PDF
2. NSB is not correctly referenced in line 14 @nemenman2002entropy
3. Please provide guidelines on how to contribute and report issues, ideally in the README files.

Best,
Yiyu
