+++ paper/paper.bib
@@ -77,7 +77,7 @@ doi = {10.1023/A:1026096204727}
 }
 
 
-@inproceedings{10.5555/2980539.2980601,
+@inproceedings{nemenman2002entropy,
   author = {Nemenman, Ilya and Shafee, Fariel and Bialek, William},
   title = {Entropy and inference, revisited},
   year = {2001},
