+++ src/Estimators/Bayesian/nsb.jl
@@ -26,60 +26,60 @@ that ``N/K → 0``, which we set to be ``N/K < 0.1`` by default
 [Asymptotic NSB estimator](https://arxiv.org/pdf/physics/0306063.pdf) (equations 11 and 12)
 """
 function ansb(data::CountData; undersampled::Float64=0.1, std_dev=false, verbose=false)
+  if verbose
+    rd = ratio(data)
+    if rd > undersampled
+      @warn("data is not sufficiently undersampled $rd, so calculation may diverge...")
+    end
+  end
+
+  Δ = coincidences(data)
+  if iszero(Δ)
     if verbose
-        rd = ratio(data)
-        if rd > undersampled
-            @warn("data is not sufficiently undersampled $rd, so calculation may diverge...")
-        end
+      @warn("no coincidences")
     end
+    return NaN
+  end
 
-    Δ = coincidences(data)
-    if iszero(Δ)
-        if verbose
-            @warn("no coincidences")
-        end
-        return NaN
-    end
-
-    if std_dev
-        return (γ - log(2)) + 2 * log(data.N) - digamma(Δ), sqrt(trigamma(Δ))
-    else
-        return (γ - log(2)) + 2 * log(data.N) - digamma(Δ)
-    end
+  if std_dev
+    return (γ - log(2)) + 2 * log(data.N) - digamma(Δ), sqrt(trigamma(Δ))
+  else
+    return (γ - log(2)) + 2 * log(data.N) - digamma(Δ)
+  end
 end
 
 function dlogrho(K0, K1, N)
-    # equation 15 from Inference of Entropies of Discrete Random Variables with Unknown Cardinalities,
-    # rearranged to make solving for 0 easier
-    K1 / K0 - digamma(K0 + N) + digamma(K0)
+  # equation 15 from Inference of Entropies of Discrete Random Variables with Unknown Cardinalities,
+  # rearranged to make solving for 0 easier
+  K1 / K0 - digamma(K0 + N) + digamma(K0)
 end
 
 function find_extremum_log_rho(K::Int64, N::Float64)
-    func(K0) = dlogrho(K0, K, N)
+  func(K0) = dlogrho(K0, K, N)
 
-    # z = find_zero(func, 1)
-    # println("$z, $K, $N")
-    # return z
-    return find_zero(func, 1)
+  # z = find_zero(func, 1)
+  # println("$z, $K, $N")
+  # return z
+  return find_zero(func, 1)
 end
 
 function neg_log_rho(data::CountData, β, K::Int64)
-    # equation 8 from Inference of Entropies of Discrete Random Variables with Unknown Cardinalities,
-    # rearranged to take logarithm (to avoid overflow)
-    κ = K * β
+  # equation 8 from Inference of Entropies of Discrete Random Variables with Unknown Cardinalities,
+  # rearranged to take logarithm (to avoid overflow)
+  κ = K * β
 
-    return -(
-        (loggamma(κ) - loggamma(data.N + κ)) +
-        (sum(x[2] * (loggamma(x[1] + β) - loggamma(β)) for x in eachcol(data.multiplicities))))
+  return -(
+    (loggamma(κ) - loggamma(data.N + κ)) +
+    (sum(x[2] * (loggamma(x[1] + β) - loggamma(β)) for x in eachcol(data.multiplicities))))
 end
 
 function find_l0(K, data::CountData)
-    neg_log_rho(data, find_extremum_log_rho(K, data.N) / K, K)
[…]
