+++ .github/workflows/CI.yml
@@ -17,7 +17,7 @@ jobs:
       matrix:
         version:
           - '1.8' # Replace this with the minimum Julia version that your package supports.
-          - '1' # Leave this line unchanged. '1' will automatically expand to the latest stable 1.x release of Julia.
+          - '1.10' # Leave this line unchanged. '1' will automatically expand to the latest stable 1.x release of Julia.
         os:
           - ubuntu-latest
         arch:
