+++ .github/workflows/CI.yml
@@ -6,7 +6,7 @@ on:
   push:
     branches:
       - main
-    tags: '*'
+    tags: "*"
 
 jobs:
   test:
@@ -16,8 +16,8 @@ jobs:
       fail-fast: false
       matrix:
         version:
-          - '1.8' # Replace this with the minimum Julia version that your package supports.
-          - '1.10' # Leave this line unchanged. '1' will automatically expand to the latest stable 1.x release of Julia.
+          - "1.8" # Replace this with the minimum Julia version that your package supports.
+          - "1.10" # Leave this line unchanged. '1' will automatically expand to the latest stable 1.x release of Julia.
         os:
           - ubuntu-latest
         arch:
@@ -58,7 +58,7 @@ jobs:
       - uses: actions/checkout@v2
       - uses: julia-actions/setup-julia@v1
         with:
-          version: '1'
+          version: "1.10"
       - run: |
           git config --global user.name name
           git config --global user.email email
@@ -82,8 +82,8 @@ jobs:
 #       - uses: julia-actions/setup-julia@v1
 #         with:
 #           version: '1.6'
-      # - uses: julia-actions/cache@v1
-      # - name: Install dependencies
+# - uses: julia-actions/cache@v1
+# - name: Install dependencies
 #         run: julia --project=docs/ -e 'using Pkg; Pkg.develop(PackageSpec(path=pwd())); Pkg.instantiate()'
 #       - name: Build and deploy
 #         env:
