### 2024-10-14 kellino/DiscreteEntropy.jl#12 @nluetts
Comments to paper (JOSS review)
Dear @kellino,

I am opening this in response to the JOSS review https://github.com/openjournals/joss-reviews/issues/7334.

The following is a list of my comments to your paper draft. It is mainly cosmetics, but in two places I think the paper would benefit from some further detail:

**Line 9:**

In the source code, change Equation 1 to:

```
$$H(X) = - \sum_{x \in X} p(x) \log (p(x))$$
```

(remove backslash after `sum`) then LaTeX should render it correctly (you can test it with the Docker image provided by JOSS, https://joss.readthedocs.io/en/latest/paper.html#docker).

**Line 22:**

Reference to Blackwell and Böhme, replace reference in source code to

```
[@bohme:fse:2020;@blackwell2023hyperfuzzing]
```

(delimited by semi-colon), to make the reference parse correctly.

**Line 17:** "Each implementation had a different precision, ..."

Do you mean like 32 vs. 64 bit vs. big float numbers?
If so, why is this a problem?

**Line 23:** Introduce the term "KL divergence".

**Line 39 to 42:** Add references to other packages and tell what
they (do not) cover.

**Line 41:** Remove "those", capitalize "Bub"?

**Line 51:** Change "making it both" to "making them both"

**Code example:**

Delete lines 62 and and 63 and move lines 65 and 66 above the comment
in lines 60 and 61 (62 and 63 are redundant).

**Line 77 to 80:**

Change `title={...}` in source code `paper.bib` to:

```
title = {Nonparametric estimation of {S}hannon's diversity index when there are unseen species in sample.},
```

You can force LaTeX to write "Shannon" with capital S by wrapping
it in curly braces. You can do the same for the reference to Hausser & Strimmer ("James-Stein").

**Line 84 to 87:** Should these references have the same journal information
as Blackwell (arXiv)?

**Line 88:** Delete "Copyright"
