### 2024-10-14 kellino/DiscreteEntropy.jl#13 @nluetts
Suggestions for improving documentation (JOSS review)
Here are some suggestions on how to improve the docs, again with reference to the JOSS review https://github.com/openjournals/joss-reviews/issues/7334,

This is not yet complete, but as far as I got today.

I added "[X]" to some of these points with X = A, B, C, ... so you can easily
reference them if you want to comment. The headings just follow the structure of your documentation.

# Overview

## Summary

[A] After "DiscreteEntropy implements a large collection of entropy
estimators.", add a list of all implemented estimators, so the user can get a
quick, central overview (otherwise one has to scroll through the "Estimators"
section where there is no table of contents that lists all estimators). Also,
list the measures that one can calculate with your package (cross entropy, KL
divergence, ...,  mutual information, conditional entropy, etc.) and link to
their specific API documentation.

In general, the documentation is very good in terms of how comprehensive
it is, but since most of it is API documentation, it is lacking guidance
beyond the "Overview" page.

## Installing Discrete Entropy

Add spaces before "higher are supported. ..."
to make the markdown list render correctly:

```md
1. If you have not done so already, install [Julia](https://julialang.org/downloads/). Julia 1.8 and
   higher are supported. Nightly is not (yet) supported.
```

[B] Mention the current limit to Julia "<=1.10".

## Basic usage

[C] Add a new subsection (##) "Contributing" and link to `CONTRIBUTING.md`.
This would be a good bonus:
If a contributor would want to add their own estimator, how would that work?
(What is the interface, or what functions would need to be implemented?)

# Data

## `CountData`

[D] Why is the total number of samples `CountData.N` represented with a
floating point number? Is it not always an integer? Is the same true for the
multiplicity matrix?

Below the matrix, change "We have 1 bin with ~~a~~ 4 elements" with
"We have 1 bin with 4 elements". This one "a" can make it really
confusing :-)

## `from_data`

"A 0 value in the ~~histgram~~ histogram ..."

# Estimate_H

## `estimate_h`

"Return the estimate**d** ~~in nats of~~ Shannon entropy of `data` **in nats**
using `estimator`."

"Note

While most calls to estimate_h take a CountData struct, this is not true for
every estimator, especially those that work directly over samples, or need the
original structure of the histogram.

For a complete list of methods of the function, try

```
julia> methods(estimate_h)
```
"

[E] The call to `methods` returns a long list of method calls that take a CountData
object and only a single call that takes a SampleVecor:

```
...
[25] estimate_h(data::SampleVector, estimator::Type{T}, ::Type{Bootstrap}; seed, reps, concentration) where T<:AbstractEstimator
...
```

However, in this case every estimator seems to be allowed
(`T<:AbstractEstimator`), so it does not get clear at this point
which estimators require which kind of data.

[F] Put

"This function is a wrapper indended to make using the libary easier. For finer
control over some of the estimators, it is advisable to call them directly,
rather than through this function."

right under

"Return the estimate in nats of Shannon entropy of data using estimator."

(or the revised version).

# Estimators

[G] I would advise putting the API reference for the abstract base types
at the end, since they are mainly of interest for contributors.
On the other hand, you should put the types for the concrete estimators here
(and in the overview as a list). Since you advertised `estimate_h` which needs
them as inputs earlier in the docs, you should provide these types now, since
one cannot always guess from the function names how the corresponding types are
called exactly (e.g. LaPlace --- is this  really how it should be written, by
the way?).

## `grassberger`

The link in

"Equation 13 from Finite sample corrections to entropy and dimension estimate"

yields a 404.

## `schurmann`

"~~This~~ There is no one ideal ..."

## `schurmann_generalized`

The link underneath the function signature should go to the end and
have "External Links" above it.

[H] "Computes the generalised Schurmann entropy estimation, given a countvector
data and a vector of xi values."

This says "vector of xi values" but the function signature above it says
`xis::Distribution`.

## `bub`

It would be good to explain the optional keyword arguments.

The external link goes to a blank page saying:

"Your session has timed out. Please go back to the article page and click the PDF link again."

(The URL contains a token that is invalid.)

## `chao_wang_jost`

The external link does not open for me.

## `unseen`

The docs do not render correctly for this one, seems a "`" is missing.

The external link requires an google account and access permission.

In general it would be good to give dois of the cited papers,
so people can find them even when the links are not working.

## `minimax`

The math is not rendered correctly.

## `nsb`

The function signature does not include type information.

## `ansb`

Should say what keyword argument `undersampled` is (N/K?).

## `pym`

[I] This function does not conform to the standard API.
What are the arguments `mm` and `icts`? Why does it not
take CountData?
