+++ paper/paper.bib
@@ -37,7 +37,7 @@ eprint = {https://doi.org/10.1137/1104033},
   interhash = {86f621d9d6f9f159448f768d792d4511},
   intrahash = {4c23fea472f6e75c0964badd83883d77},
   keywords = {bayesianinference book informationtheory neuralnetworks patternrecognition probabilitytheory},
-  publisher = {Copyright Cambridge University Press},
+  publisher = {Cambridge University Press},
   timestamp = {2007-05-24T14:43:04.000+0200},
   title = {Information Theory, Inference, and Learning Algorithms},
   doi = {10.1109/tit.2004.834752},
@@ -70,7 +70,7 @@ author = {Chao, Anne and Shen, Tsung-Jen},
 year = {2003},
 month = {12},
 pages = {429-443},
-title = {Chao A, Shen T-J.. Nonparametric estimation of Shannon's diversity index when there are unseen species in sample. Environ Ecol Stat 10: 429-443},
+title = {Nonparametric estimation of {S}hannon's diversity index when there are unseen species in sample. Environ Ecol Stat 10: 429-443},
 volume = {10},
 journal = {Environmental and Ecological Statistics},
 doi = {10.1023/A:1026096204727}
@@ -108,11 +108,31 @@ doi = {10.1023/A:1026096204727}
 }
 
 @misc{hausser2009entropy,
-      title={Entropy inference and the James-Stein estimator, with application to nonlinear gene association networks},
+      title={Entropy inference and the {J}ames-{S}tein estimator, with application to nonlinear gene association networks},
       author={Jean Hausser and Korbinian Strimmer},
       year={2009},
       eprint={0811.3579},
       archivePrefix={arXiv},
       primaryClass={stat.ML}
 }
+
+@book{Cover2006,
+  added-at = {2009-04-20T21:27:16.000+0200},
+  at = {2008-03-31 06:17:47},
+  author = {Cover, Thomas M. and Thomas, Joy A.},
+  biburl = {https://www.bibsonomy.org/bibtex/22e9bfa879286689a14feb55b69d326c1/ywhuang},
+  howpublished = {Hardcover},
+  id = {1877660},
+  interhash = {87ae368776946bf7a71ee476e81a2191},
+  intrahash = {2e9bfa879286689a14feb55b69d326c1},
+  isbn = {0471241954},
+  keywords = {information-theory book},
+  month = {July},
+  priority = {0},
+  publisher = {Wiley-Interscience},
+  timestamp = {2009-04-20T21:27:16.000+0200},
+  title = {Elements of Information Theory 2nd Edition (Wiley Series in Telecommunications and Signal Processing)},
+  year = 2006
+}
+
 
