### 2024-10-15 82a15cb
added extra info and made edits
