+++ paper/paper.md
@@ -39,13 +39,13 @@ This difficulty has led to a large number of improved estimators. [@Rodriguez202
 evaluate 18 different estimators, among which are _Grassberger_ [@grassberger2008entropy],
 _Chao Shen_ [@chaoshen] , _NSB_ [@nemenman2002entropy], _Zhang_ [@zhang] and _James-Stein_ [@hausser2009entropy].
 These estimators were scattered across 3 different programming languages
-and 7 different libraries. Some of these estimators are hard to find or poorly maintained. Each implementation had
-a different precision, making comparison of estimations difficult.
+and 7 different libraries. Some of these estimators are hard to find or poorly maintained. Each implementation
+calculates and reports entropy to a different number of significant digits, which can lead to difficulties in comparison.
 
 If one can estimate entropy more accurately, then one can also estimate mutual information more accurately. There
 are numerous, cross-domain, applications for entropy and mutual information, such as in telecommunications,
-machine learning [@MacKay2003] and software engineering [@bohme:fse:2020, @blackwell2023hyperfuzzing]. `DiscreteEntropy.jl` makes
-it easy to apply different estimators to the problem of mutual information, cross entropy and KL divergence, amongst other
+machine learning [@MacKay2003] and software engineering [@bohme:fse:2020;@blackwell2023hyperfuzzing]. `DiscreteEntropy.jl` makes
+it easy to apply different estimators to the problem of mutual information, cross entropy and Kullback–Leibler divergence[@Cover2006], amongst other
 measures.
 
 `DiscreteEntropy.jl` provides a comprehensive collection of popular entropy estimators and utilities for working with other Shannon measures.
@@ -63,7 +63,15 @@ in [@Rodriguez2021EntropyEst] and a number of estimators which were not consider
 Shannon measures for their research, or those who want to research entropy estimation directly.
 
 There is no other open-source software package known to us, in any language, with similiar features or similiar breadth of
-estimators. The estimators here were mostly implemented from the original papers, those some (such as Bub and Unseen) are
+estimators. The R [entropy](https://cran.r-project.org/web/packages/entropy/index.html) package covers many basic estimators, such as
+the maximum likelihood, Miller-Madow, Chao Shen and many bayesian estimators.
+The [PYM](https://github.com/pillowlab/PYMentropy/?tab=readme-ov-file), [BUB](http://www.stat.columbia.edu/~liam/research/code/BUBfunc.m) and
+[Unseen](http://theory.stanford.edu/~valiant/papers/nips_full.pdf)
+estimators are found only in the author's Matlab implementions. Other estimators, such as Zhang and Grassberger, can be found in the Python
+[Entropart](https://ericmarcon.r-universe.dev/entropart).
+The NSB estimator exists in multiple different versions, in [C++, Matlab](https://sourceforge.net/projects/nsb-entropy/)
+and [Python](https://github.com/simomarsili/ndd).
+The estimators here were mostly implemented from the original papers, though some (such as BUB and Unseen) are
 idiomatic ports of original Matlab code.
 
 `DiscreteEntropy.jl` is a fast, simple to use, library that fills a gap between the scattered implementations available online.
@@ -74,27 +82,32 @@ It ensures type safety throughout, even preventing confusion between vectors of 
 `DiscreteEntropy.jl` allows the user to call each estimator directly, or to use a helper function `estimate_h`.
 The `estimate_h` function is the easiest entry to the library. This function takes a `CountData` object, which
 can be constructed from a vector using either `from_data`, `from_counts` or `from_samples`. Both `from_data` and
[…]
