### 2024-10-15 fcbf70c
fix: remove backslash from equation
