### 2024-10-15 kellino/DiscreteEntropy.jl#12 @kellino
Hi @nluetts I think I've addressed all the (excellent) comments you suggested. I'll leave this open until you've signed off on the contents.
