### 2024-10-15 265f7f8
fix: return NaN for no coincidences
