### 2024-10-15 a3e2bab
fix: remove remove_zeros from samples
