+++ src/Core/countdata.jl
@@ -58,9 +58,9 @@ CountData is not expected to be called directly, nor is it advised to directly m
 [`from_samples`](@ref) instead.
 """
 mutable struct CountData
-    multiplicities::Matrix{Float64}
-    N::Float64
-    K::Int64
+  multiplicities::Matrix{Float64}
+  N::Float64
+  K::Int64
 end
 
 Base.:(==)(x::CountData, y::CountData) = Base.:(==)(x.multiplicities, y.multiplicities) && x.N == y.N && x.K == y.K
@@ -68,12 +68,12 @@ Base.copy(x::CountData) = CountData(x.multiplicities, x.N, x.K)
 Base.Broadcast.broadcastable(q::CountData) = Ref(q)
 
 function empty_countdata()
-    x::Matrix{Float64} = [;;]
-    CountData(x, 0.0, 0)
+  x::Matrix{Float64} = [;;]
+  CountData(x, 0.0, 0)
 end
 
 function Base.hash(g::CountData, h::UInt)
-    hash(g.multiplicities, hash(g.K, hash(g.N, h)))
+  hash(g.multiplicities, hash(g.K, hash(g.N, h)))
 end
 
 @doc raw"""
@@ -81,7 +81,7 @@ end
 Return the bins (top row) of x.multiplicities
 """
 function bins(x::CountData)
-    return x.multiplicities[1,:]
+  return x.multiplicities[1, :]
 end
 
 @doc raw"""
@@ -89,43 +89,43 @@ end
 Return the bin multiplicities (bottom row) of x.multiplicities
 """
 function multiplicities(x::CountData)
-    return x.multiplicities[2,:]
+  return x.multiplicities[2, :]
 end
 
 function find_col(data::CountData, target)
-    for x in eachcol(data.multiplicities)
-        if x[1] == target
-            return x
-        end
+  for x in eachcol(data.multiplicities)
+    if x[1] == target
+      return x
     end
+  end
 end
 
 function singletons(data::CountData)
-    find_col(data, 1)
+  find_col(data, 1)
 end
 
 function doubletons(data::CountData)
-    find_col(data, 2)
+  find_col(data, 2)
 end
 
 function coincidences(data::CountData)
-    data.N - data.K
+  data.N - data.K
 end
 
 function ratio(data::CountData)::Float64
-    coincidences(data) / data.N
+  coincidences(data) / data.N
 end
 
 function _from_counts(counts::AbstractVector{T}, zeros) where {T<:Real}
-    map = countmap(counts)
-    if zeros
-        delete!(map, 0)
-    end
-    x1 = collect(keys(map))
-    x2 = collect(values(map))
-    mm = [x1 x2]
+  map = countmap(counts)
+  if zeros
+    delete!(map, 0)
+  end
+  x1 = collect(keys(map))
+  x2 = collect(values(map))
+  mm = [x1 x2]
 
-    return CountData(mm', dot(x1, x2), sum(mm[:, 2]))
+  return CountData(mm', dot(x1, x2), sum(mm[:, 2]))
 end
 
 @doc raw"""
@@ -136,47 +136,46 @@ Return a [`CountData`](@ref) object from a vector or CountVector. Many estimator
 cannot handle a histogram with a 0 value bin, so there are filtered out unless remove_zeros is set to false.
 """
 function from_counts(counts::CountVector, remove_zeros::Bool)
-    if isempty(counts)
-        @warn("returning empty CountData object")
-        return empty_countdata()
-    end
-    _from_counts(counts.values, remove_zeros)
+  if isempty(counts)
+    @warn("returning empty CountData object")
+    return empty_countdata()
+  end
+  _from_counts(counts.values, remove_zeros)
 end
 
 function from_counts(counts::AbstractVector; remove_zeros::Bool=true)
[…]
