+++ src/Estimators/Frequentist/bub.jl
@@ -21,149 +21,150 @@ where h is the estimation of Shannon entropy in `nats` and MM is the upper bound
 [Estimation of Entropy and Mutual Information](https://watermark.silverchair.com/089976603321780272.pdf?token=AQECAHi208BE49Ooan9kkhW_Ercy7Dm3ZL_9Cf3qfKAc485ysgAAA1wwggNYBgkqhkiG9w0BBwagggNJMIIDRQIBADCCAz4GCSqGSIb3DQEHATAeBglghkgBZQMEAS4wEQQM0ehw1kCaNd562GwpAgEQgIIDD_n4g4aLZda9boTVZsgQpyahxo-C4fFpEfzOUj_-iZ1gJh12HjkrpqIqrcBW6r18YwtMei4RRCZu90KuoxLqfvH7P5vdOVfYM-TsChkmzKWl5ajUZADxogTOqKFYMWjovXlOTJMvj8Nj484WOxceGzLwxs-xgE5qawaIpeKR4qqIR_AhmGpKrogKHJ7SHy2feKSCzpw0mdZcAV1BFmSXPiJe4djQ3kGlEEuQLVCXfV3ZzasGUrZ-45rPBsLfn9OX2-qDnGtsmmpvt8AuMe7znggoFMTUQE1B4JL79auno5RUSfu7bXGA0vBOsBopTo86c2ENkayliPTFXDExVp_QB75zmLyBtNKE0it8brZr4HAyDfGQyY956x8a6xtVS1vtZrio9AoG3Jfl52m8heXmvsQr4M51YIBMKx5bP0ext9uaBAMwwW3XEIKlIl22EI50iKsGVV-N8LWCuyrLR9LAAo-raJAOn1mbg268rtvpfGA9_sqHRc7Anal8YgJABRL_b6f7_xzT2tclyaRa60l9-9m3l2WtQpYd7UyrhPrN5-7IYGaGyKZbek4mDys4KwyqIRZDkpcgxSuHxXUZO7jbu1e5ek9Tg4RAGSAz1901aPj6PsF3ttsIeosrfAkK4c8xyrjHBIkxZO-4zBurhgDjGh3Yeo787FaN4j1bdsfTFLUC7cXghxkDMQzO4l_gunR1J7PVnRbHGqdkvZwTveP_xDRRex-D_y2jJ3r7cemZI-WmjT-NIJSflYyI9KUgqQ6y_6rTE696ttVkbrgJ1Sh95J_ISQvLzM4AjUDwCjFqpZxGvqTmK3B4MRbWlriC3QVF_wMfW5-CM2robw3n7HjlEpHDU85k5CYvPrvZG32OVU5Y9wI_PZTe94o2KuoYaC9cShqyk90lZmSN4gBz3bMgyWpGwZLy1-U84xpMphAzUHqsDV4wNQwJhbSRSO6d7G07DrfBm5yQnUXatJOTNZlZrM9UWu5e2pFCVjt79onv1TYbYA6USTsoewkHvlOaOFrXs2P07ESqLC1CIA2HcDYEkw)
 """
 function bub(data::CountData; k_max=11, truncate=false, lambda=0.0)
-    if k_max > data.N
-        k_max = floor(Int, data.N)
-    end
-    if data.N < 20.0
-        under(data)
-    else
-        over(data, k_max, truncate, lambda_0=lambda)
-    end
+  if k_max > data.N
+    k_max = floor(Int, data.N)
+  end
+  if data.N < 20.0
+    under(data)
+  else
+    over(data, k_max, truncate, lambda_0=lambda)
+  end
 end
 
 function under_ub(N, K, a, mesh)
-    p, P = get_mesh(N, mesh)
-    Pn = [dot(a, x) for x in eachcol(P)]
[…]
