### 2024-10-18 #7334 @niyiyu
Thanks for the updates. The equation is still not rendering thought. Other than that I don't have any concerns and will recommend for accepting for publishing.  @Nikoleta-v3
