### 2024-10-18 #7334 @kellino
Thank you @niyiyu. The equation should be rendering correctly now, as of the most recent commit https://github.com/kellino/DiscreteEntropy.jl/commit/ed5577923c3b5659a881701c0deb1329fa86b0b2
