### 2024-10-18 kellino/DiscreteEntropy.jl#13 @nluetts
Here is the rest of my suggestions.

@kellino I am pinging you here, in case you did not see this issue.

# Mixed Estimators

## `pert`

I would not use verbatim for explaining the parameter a,b,c here, because
usually it is only used for code examples.

[J] Explain the arguments. What is swapped with `estimator`, a or b or c?
What about `e1` and `e2`, a and c? Why can one not choose a, b and c
freely?

## `jackknife`

What is `corrected`?

## `bayesian_bootstrap`

Put keyword arguments `seed` and `concentration` in description in backticks.

The link is behind a paywall. Is there no paper that describes how
Bayesian bootstrapping works?

# Utility Functions

## `logx`

[K] I do not understand why returning 0 instead of -Inf (like the
standard Julia function) is a good idea. You use this function in your
code, so perhaps you can explain why this is fine in the context of
entropy estimation?

## `bins`

Use verbatim for `x.multiplicities`

## `from_csv`

~~*CSV.File()~~ → `CSV.File()`

You can pass all further keyword arguments on to `CSV.File()` by changing the
function signature to

```julia
from_csv(file::String, field, ::Type{T}; remove_zeros=false, header=nothing, kw...)
```

and swapping `for row in CSV.File(file; header=header)` in the source code
with `for row in CSV.File(file; header=header, kw...)`

Then, users will have more options before falling back to using CSV.jl
directly.

# Divergence and Distance

## `cross_entropy`

[L] In the function signature, should it really be `{T<:MaximumLikelihood}`?
Looking at the source code, this should be a typo.

## `kl_divergence`

"~~Both distributions needs to be defined over the same space, so length(p) == length(q).~~"

Use the same statement as in the description of `cross_entropy`: "P and Q must be the same length."

"If the distributions are not over the same space **(their cross entropy is < 0)**, ~~then it~~the function return**s** Inf."

(cross entropy < 0 according to source code, I find this easier to understand)

## `jensen_shannon_divergence`

[N] The description says "P and Q, as represented by their histograms."
Why does the function take `AbstractVector`s and not `CountVector`s?

## `jensen_shannon_distance`

[O] Arguments are called `P` and `Q` like for `cross_entropy` and
`kl_divergence`, where they are of type `CountData`, but here they
are `AbstractVector`s. Should they be `CountVector`s, like for
`jensen_shannon_divergence`?

## `uncertainty_coefficient`

[P] What is `joint` in the equations in the description? *I(X;Y)*?

# Mutual Information and Conditional Information

## `mutual_information`

Description has some typo, does not render correctly:

"using Optim: estimatemaxstep mutualinformation(joint::Matrix{I}, estimator::Type{T}) where {T<:AbstractEstimator, I<:Real}"

(looks like a surplus or misplaced `using Optim: estimate_maxstep` in the docstring)

[Q] What is `XY` and where would it come from? I think this one
would benefit from an example.

## `conditional_entropy`

Explain the keyword arguments.

"and estimator ~~*estimator*~~`estimator`" (verbatim)
