### 2024-10-18 kellino/DiscreteEntropy.jl#11 @kellino
The original author's implementation of NSB returns NaN when there are no coincidences in the data, the python implementation does not. I have opted to go with the author's decision and return NaN. Unseen had a bug which has now been fixed. I've been unable to find any errors in the remaining estimators by comparing against other implementations, so I'm closing here, though the question remains open as to why Rodriguez reports some differing results.
