+++ src/Core/utils.jl
@@ -6,33 +6,42 @@ using StatsBase: countmap;
 @doc raw"""
     logx(x)::Float64
 Returns natural logarithm of x, or 0.0 if x is zero
+
+Setting logx(0) to 0 is a standard trick used when estimating entropy in order to
+avoid infinities when calculating the average surprisal.
 """
 function logx(x)
-    if iszero(x)
-        return zero(x)
-    end
-    return log(x)
+  if iszero(x)
+    return zero(x)
+  end
+  return log(x)
 end
 
 @doc raw"""
     xlogx(x::Float64)
 Returns `x * log(x)` for `x > 0`, or 0.0 if x is zero
+
+Setting xlogx(0) to 0 is a standard trick used when estimating entropy in order to
+avoid infinities when calculating the average surprisal.
 """
 function xlogx(x)
-    return x * logx(x)
+  return x * logx(x)
 end
 
 @doc raw"""
      xFx(f::Function, x)
 
 Returns `x * f(x)` for `x > 0`, or 0.0 if x is zero
+
+Setting xFx(0) to 0 is a standard trick used when estimating entropy in order to
+avoid infinities when calculating the average surprisal.
 """
 function xFx(f::Function, x)
-    if iszero(x)
-        return zero(x)
-    else
-        return x * f(x)
-    end
+  if iszero(x)
+    return zero(x)
+  else
+    return x * f(x)
+  end
 end
 
 @doc raw"""
@@ -40,7 +49,7 @@ end
 Return ``\frac{h}{\log(2)}`` where h is in nats
 """
 function to_bits(h::Float64)
-    return h / log(2)
+  return h / log(2)
 end
 
 @doc raw"""
@@ -48,7 +57,7 @@ end
 Return ``\frac{h}{log(10)}`` where `h` is in nats
 """
 function to_bans(h::Float64)
-    return h / log(10)
+  return h / log(10)
 end
 
 @doc raw"""
@@ -58,7 +67,7 @@ Return ``\log \Gamma(x + δx) - \log Γ(x)``
 
 """
 function gammalndiff(x::Float64, dx::Float64)
-    return loggamma(x + dx) - loggamma(x)
+  return loggamma(x + dx) - loggamma(x)
 end
 
 @doc raw"""
@@ -70,34 +79,34 @@ If normalised = true, return as probability distribution.
 
 """
 function marginal_counts(joint::Matrix, dim; normalise=false)
-    p = nothing
-    if dim == 1
-        p = [sum(x) for x in eachcol(joint)]
-    end
-    if dim == 2
-        p = [sum(x) for x in eachrow(joint)]
-    end
+  p = nothing
+  if dim == 1
+    p = [sum(x) for x in eachcol(joint)]
+  end
+  if dim == 2
+    p = [sum(x) for x in eachrow(joint)]
+  end
 
-    if p === nothing
-        return p
-    else
-        if normalise
-            return p ./ sum(p)
-        end
-    end
-
+  if p === nothing
     return p
+  else
+    if normalise
+      return p ./ sum(p)
+    end
+  end
+
+  return p
 
 end
 
 function logspace(start, stop, steps::Integer)
-    return 10 .^ range(start, stop, length=steps)
+  return 10 .^ range(start, stop, length=steps)
 end
 
 function update_dict!(d, k; v=1)
-    if haskey(d, k)
-        d[k] += v
-    else
-        d[k] = v
-    end
+  if haskey(d, k)
+    d[k] += v
+  else
+    d[k] = v
+  end
 end
