+++ src/Estimators/Frequentist/frequentist.jl
@@ -18,9 +18,9 @@ or equivalently
 ```
 """
 function maximum_likelihood(data::CountData)
-    log(data.N) -
-    (1.0 / data.N) *
-    sum(xlogx(x[1]) * x[2] for x in eachcol(data.multiplicities))
+  log(data.N) -
+  (1.0 / data.N) *
+  sum(xlogx(x[1]) * x[2] for x in eachcol(data.multiplicities))
 end
 
 
@@ -32,13 +32,14 @@ end
 Compute the *jackknifed* [`maximum_likelihood`](@ref) estimate of data and the variance of the
 jackknifing (not the variance of the estimator itself).
 
-If corrected is true, then the variance is scaled with n-1, else it is scaled with n
+If `corrected` is true, then the variance is scaled with $data.N-1$, else it is scaled with $data.N$. `corrected`
+has no effect on the entropy estimation.
 
 # External Links
 [Estimation of the size of a closed population when capture probabilities vary among animals](https://academic.oup.com/biomet/article/65/3/625/234287)
 """
 function jackknife_mle(data::CountData; corrected=false)
-    return jackknife(data, MaximumLikelihood, corrected=corrected)
+  return jackknife(data, MaximumLikelihood, corrected=corrected)
 end
 
 
@@ -55,7 +56,7 @@ on the total number of samples seen (N) and the support size (K).
 ```
 """
 function miller_madow(data::CountData)
-    return maximum_likelihood(data) + ((data.K - 1.0) / (2.0 * data.N))
+  return maximum_likelihood(data) + ((data.K - 1.0) / (2.0 * data.N))
 end
 
 
@@ -70,44 +71,36 @@ Compute the Grassberger (1988) estimation of Shannon entropy of `data` in nats
 \hat{H}_{\tiny{Gr88}} = \sum_i \frac{h_i}{H} \left(\log(N) - \psi(h_i) - \frac{(-1)^{h_i}}{n_i + 1}  \right)
 ```
 Equation 13 from
-[Finite sample corrections to entropy and dimension estimate](https://www.academia.edu/download/47091312/0375-9601_2888_2990193-420160707-16069-1k3ppo7.pdf)
+[Finite sample corrections to entropy and dimension estimate](https://www.academia.edu/10831948/Finite_sample_corrections_to_entropy_and_dimension_estimates)
 """
 function grassberger(data::CountData)
-    log_n = log(data.N)
-    sum((x[1] / data.N * (log_n - digamma(x[1]) - (-1)^x[1] / (x[1] + 1))) * x[2] for x in eachcol(data.multiplicities))
+  log_n = log(data.N)
+  sum((x[1] / data.N * (log_n - digamma(x[1]) - (-1)^x[1] / (x[1] + 1))) * x[2] for x in eachcol(data.multiplicities))
 end
 
-# function grassberger2003(data::CountData)
-#     schurmann(data, 1.0)
-# end
-
-
-# Schurmann Estimator
-
 @doc raw"""
-    schurmann(data::CountData, ξ::Float64 = ℯ^(-1/2))
+     schurmann(data::CountData, ξ::Float64 = ℯ^(-1/2))
 
 Compute the Schurmann estimate of Shannon entropy of `data` in nats.
 
 ```math
 \hat{H}_{SHU} = \psi(N) - \frac{1}{N} \sum_{i=1}^{K} \, h_i \left( \psi(h_i) + (-1)^{h_i} ∫_0^{\frac{1}{\xi} - 1} \frac{t^{h_i}-1}{1+t}dt \right)
-
 ```
[…]

[…]
