### 2024-10-18 kellino/DiscreteEntropy.jl#13 @kellino
Hi @nluetts, this is an extraordinary effort on your part. Thank you very much indeed! I have pushed (to paper and main) and hope that I've covered every point that you've raised directly in the docs. I'm going to close this issue now, but please reopen if you think I've missed something or not done something very well.
