+++ paper/paper.bib
@@ -1,0 +1,138 @@
+@article{basharin,
+author = {Basharin, G. P.},
+title = {On a Statistical Estimate for the Entropy of a Sequence of Independent Random Variables},
+journal = {Theory of Probability \& Its Applications},
+volume = {4},
+number = {3},
+pages = {333-336},
+year = {1959},
+doi = {10.1137/1104033},
+URL = {https://doi.org/10.1137/1104033},
+eprint = {https://doi.org/10.1137/1104033},
+}
+
+
+@INPROCEEDINGS{bohme:fse:2020,
+ author = {B{\"o}hme, Marcel and Man{\`e}s, Valentin and Cha, Sang Kil},
+ title = {Boosting Fuzzer Efficiency: An Information Theoretic Perspective},
+ booktitle = {Proceedings of the 14th Joint meeting of the European Software Engineering Conference and the ACM SIGSOFT Symposium on the Foundations of Software Engineering},
+ series = {ESEC/FSE},
+ year = {2020},
+ pages = {970-981},
+ numpages = {12},
+ doi = {10.1145/3368089.3409748}
+}
+
+@article{blackwell2023hyperfuzzing,
+  title={Hyperfuzzing: black-box security hypertesting with a grey-box fuzzer},
+  author={Blackwell, Daniel and Becker, Ingolf and Clark, David},
+  journal={arXiv preprint arXiv:2308.09081},
+  year={2023}
+}
+
+@book{MacKay2003,
+  added-at = {2007-05-24T14:43:04.000+0200},
+  author = {MacKay, David J. C.},
+  biburl = {https://www.bibsonomy.org/bibtex/24c23fea472f6e75c0964badd83883d77/tmalsburg},
+  interhash = {86f621d9d6f9f159448f768d792d4511},
+  intrahash = {4c23fea472f6e75c0964badd83883d77},
+  keywords = {bayesianinference book informationtheory neuralnetworks patternrecognition probabilitytheory},
+  publisher = {Cambridge University Press},
+  timestamp = {2007-05-24T14:43:04.000+0200},
+  title = {Information Theory, Inference, and Learning Algorithms},
+  doi = {10.1109/tit.2004.834752},
+  year = 2003
+}
+
+@article{Rodriguez2021EntropyEst,
+  title={Selecting an effective entropy estimator for short sequences of bits and bytes with maximum entropy},
+  author={Contreras Rodr{\'\i}guez, Lianet and Madarro-Cap{\'o}, Evaristo Jos{\'e} and Leg{\'o}n-P{\'e}rez, Carlos Miguel and Rojas, Omar and Sosa-G{\'o}mez, Guillermo},
+  journal={Entropy},
+  volume={23},
+  number={5},
+  pages={561},
+  year={2021},
+  doi={10.3390/e23050561},
+  publisher={MDPI}
+}
+
+@misc{grassberger2008entropy,
+      title={Entropy Estimates from Insufficient Samplings},
+      author={P. Grassberger},
+      year={2008},
+      eprint={physics/0307138},
+      archivePrefix={arXiv},
+      primaryClass={physics.data-an}
+}
+
+@article{chaoshen,
+author = {Chao, Anne and Shen, Tsung-Jen},
+year = {2003},
+month = {12},
+pages = {429-443},
+title = {Nonparametric estimation of {S}hannon's diversity index when there are unseen species in sample. Environ Ecol Stat 10: 429-443},
+volume = {10},
+journal = {Environmental and Ecological Statistics},
+doi = {10.1023/A:1026096204727}
+}
+
+
+@inproceedings{nemenman2002entropy,
+  author = {Nemenman, Ilya and Shafee, Fariel and Bialek, William},
[…]
