+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"est_h/#Estimate_H","page":"Estimate_H","title":"Estimate_H","text":"","category":"section"},{"location":"est_h/","page":"Estimate_H","title":"Estimate_H","text":"The main entry point for the library.","category":"page"},{"location":"est_h/","page":"Estimate_H","title":"Estimate_H","text":"estimate_h","category":"page"},{"location":"est_h/#DiscreteEntropy.estimate_h","page":"Estimate_H","title":"DiscreteEntropy.estimate_h","text":"estimate_h(data::CountData, estimator::Type{T}) where {T<:AbstractEstimator}\nestimate_h(data::CountData, ::Type{JackknifeMLE}; corrected=false)\nestimate_h(data::CountData, ::Type{Schurmann}, xi=nothing)\n\nestimate_h(data::CountVector, ::Type{SchurmannGeneralised}, xis::XiVector)\nestimate_h(data::CountData, ::Type{Bayes}, α::AbstractFloat; K=nothing)\nestimate_h(data::CountData, ::Type{NSB}; guess=false, K=nothing)\n\nReturn the estimate of Shannon entropy of data, in nats, using estimator.\n\nExample\n\njulia> import Random; Random.seed!(1);\n\njulia> X = rand(1:10, 1000)\njulia> estimate_h(from_data(X, Samples), Schurmann)\n2.3039615201251173\n\nNote\n\nWhile most calls to estimateh take a CountData struct, this is not true for every estimator. In particular `schurmanngeneralised` has a different method call\n\nestimate_h(data::CountVector, ::Type{SchurmannGeneralised}, xis::XiVector)\n\nas this works directly over this sample histogram.\n\nFor a complete list of methods of the function, try\n\njulia> methods(estimate_h)\n\nReturn the estimate of Shannon entropy of data, in nats, using estimator.\n\nThis function is a wrapper indended to make using the libary easier. For finer control over some of the estimators, it is advisable to call them directly, rather than through this function.\n\n\n\n\n\n","category":"function"},{"location":"divergence/#Divergence-and-Distance","page":"Divergence and Distance","title":"Divergence and Distance","text":"","category":"section"},{"location":"divergence/","page":"Divergence and Distance","title":"Divergence and Distance","text":"cross_entropy\nkl_divergence\njensen_shannon_divergence\njensen_shannon_distance\njeffreys_divergence\nuncertainty_coefficient","category":"page"},{"location":"divergence/#DiscreteEntropy.cross_entropy","page":"Divergence and Distance","title":"DiscreteEntropy.cross_entropy","text":" cross_entropy(P::CountVector, Q::CountVector, ::Type{T}) where {T<:AbstractEstimator}\n\nH(PQ) = - sum_x(P(x) log(Q(x)))\n\nCompute the cross entropy of P and Q, given an estimator of type T. P and Q must be the same length. Both vectors are normalised. The cross entropy of a probability distribution P with itself, is equal to its entropy, ie H(P P) = H(P).\n\nExample\n\n\njulia> P = cvector([1,2,3,4,3,2])\njulia> Q = cvector([2,5,5,4,3,4])\n\njulia> ce = cross_entropy(P, Q, MaximumLikelihood)\n1.778564897565542\n\nNote: not every estimator is currently supported.\n\n\n\n\n\n","category":"function"},{"location":"divergence/#DiscreteEntropy.kl_divergence","page":"Divergence and Distance","title":"DiscreteEntropy.kl_divergence","text":"kl_divergence(P::CountVector, Q::CountVector, estimator::Type{T}; truncate::Union{Nothing, Int} = nothing) where {T<:AbstractEstimator}\n\nD_KL(P  Q) = su
[…]
