### 2024-10-18 5b06817
a few more tests
