### 2024-10-21 kellino/DiscreteEntropy.jl#12 @nluetts
Hi @kellino

I think you should cite the other packages in the new paragraph about other estimator implementations differently.
The problem with using just the links to websites is that these external pages may change or get deleted. You should include info about the access date so that readers can try pulling the website from the internet archive, if the link does not work anymore. (Yes, readers could also look at the publication date of your paper, but having the access date right there in the paper makes it easier).
You should also check if the cited packages want a paper to be cited when you reference them. I just checked for the R package `entropy` where it says in the docs (first page, section "Descritption"):

> If you use this package please cite: Jean Hausser and Korbinian Strimmer. 2009. Entropy inference
> and the James-Stein estimator, with application to nonlinear gene association networks. J. Mach.
> Learn. Res. 10: 1469-1484. Available online from https://jmlr.csail.mit.edu/papers/v10/
> hausser09a.html.

Seems like this is the final paper of the arxiv preprint that you cite like so:

```bibtex
@misc{hausser2009entropy,
      title={Entropy inference and the {J}ames-{S}tein estimator, with application to nonlinear gene association networks},
      author={Jean Hausser and Korbinian Strimmer},
      year={2009},
      eprint={0811.3579},
      archivePrefix={arXiv},
      primaryClass={stat.ML}
}
```

so you can change this citation and use it also for `entropy`.
