### 2024-10-22 kellino/DiscreteEntropy.jl#12 @kellino
Hi @nluetts I'll need a few days to action this as I'm a big snowed under, but I'll get these improvements done over the next few days and ping you when I've finished
