### 2024-10-22 kellino/DiscreteEntropy.jl#12 @nluetts
Sure, no need to hurry :-)
