+++ paper/paper.bib
@@ -135,4 +135,64 @@ doi = {10.1023/A:1026096204727}
   year = 2006
 }
 
+@misc{hausser2009entropy,
+      title={Entropy inference and the {J}ames-{S}tein estimator, with application to nonlinear gene association networks},
+      author={Jean Hausser and Korbinian Strimmer},
+      year={2009},
+      eprint={0811.3579},
+      archivePrefix={arXiv},
+      primaryClass={stat.ML}
+}
+
+@misc{pym, 
+  title = {PYM},
+  author = {pillowlab},
+  year = 2024,
+  note = {\url{https://github.com/pillowlab/PYMentropy/?tab=readme-ov-file} [Accessed: 2024-10-22]} 
+} 
+
+@misc{bub, 
+  title = {BUB},
+  author = {Liam Paninski},
+  year = 2024,
+  note = {\url{http://www.stat.columbia.edu/~liam/research/code/BUBfunc.m} [Accessed: 2024-10-22]} 
+} 
+
+@misc{nsb, 
+  title = {NSB},
+  author = {Ilya Nemenman},
+  year = 2024,
+  note = {\url{https://sourceforge.net/projects/nsb-entropy/} [Accessed: 2024-10-22]} 
+} 
+
+@misc{ndd, 
+  title = {NDD},
+  author = {Simone Marsili},
+  year = 2024,
+  note = {\url{https://github.com/simomarsili/ndd} [Accessed: 2024-10-22]} 
+} 
+
+@article{entropart,
+  title = {{entropart}: An {R} Package to Measure and Partition Diversity},
+  author = {Eric Marcon and Bruno H{\'e}rault},
+  journal = {Journal of Statistical Software},
+  year = {2015},
+  volume = {67},
+  number = {8},
+  pages = {1--26},
+  doi = {10.18637/jss.v067.i08},
+}
+
+@inproceedings{valiant,
+ author = {Valiant, Paul and Valiant, Gregory},
+ booktitle = {Advances in Neural Information Processing Systems},
+ editor = {C.J. Burges and L. Bottou and M. Welling and Z. Ghahramani and K.Q. Weinberger},
+ pages = {},
+ publisher = {Curran Associates, Inc.},
+ title = {Estimating the Unseen: Improved Estimators for Entropy and other Properties},
+ url = {https://proceedings.neurips.cc/paper_files/paper/2013/file/53c04118df112c13a8c34b38343b9c10-Paper.pdf},
+ volume = {26},
+ year = {2013}
+}
+
 
