+++ paper/paper.bib
@@ -151,12 +151,13 @@ doi = {10.1023/A:1026096204727}
   note = {\url{https://github.com/pillowlab/PYMentropy/?tab=readme-ov-file} [Accessed: 2024-10-22]} 
 } 
 
-@misc{bub, 
-  title = {BUB},
-  author = {Liam Paninski},
-  year = 2024,
-  note = {\url{http://www.stat.columbia.edu/~liam/research/code/BUBfunc.m} [Accessed: 2024-10-22]} 
-} 
+@misc{bub,
+	author = {Liam Paninski},
+	title = {stat.columbia.edu},
+	howpublished = {\url{http://www.stat.columbia.edu/~liam/research/code/BUBfunc.m}},
+	year = {},
+	note = {[Accessed 22-10-2024]},
+}
 
 @misc{nsb, 
   title = {NSB},
