### 2024-10-22 2177e54
better bibtex format
