+++ paper/paper.bib
@@ -147,13 +147,14 @@ doi = {10.1023/A:1026096204727}
 @misc{pym, 
   title = {PYM},
   author = {pillowlab},
-  year = 2024,
-  note = {\url{https://github.com/pillowlab/PYMentropy/?tab=readme-ov-file} [Accessed: 2024-10-22]} 
+  year = {},
+  howpublished = {\url{https://github.com/pillowlab/PYMentropy/?tab=readme-ov-file}}, 
+	note = {[Accessed 22-10-2024]},
 } 
 
 @misc{bub,
 	author = {Liam Paninski},
-	title = {stat.columbia.edu},
+	title = {BUB},
 	howpublished = {\url{http://www.stat.columbia.edu/~liam/research/code/BUBfunc.m}},
 	year = {},
 	note = {[Accessed 22-10-2024]},
@@ -162,15 +163,17 @@ doi = {10.1023/A:1026096204727}
 @misc{nsb, 
   title = {NSB},
   author = {Ilya Nemenman},
-  year = 2024,
-  note = {\url{https://sourceforge.net/projects/nsb-entropy/} [Accessed: 2024-10-22]} 
+  year = {},
+  howpublished = {\url{https://sourceforge.net/projects/nsb-entropy/}}, 
+	note = {[Accessed 22-10-2024]},
 } 
 
 @misc{ndd, 
   title = {NDD},
   author = {Simone Marsili},
-  year = 2024,
-  note = {\url{https://github.com/simomarsili/ndd} [Accessed: 2024-10-22]} 
+  year = {},
+  howpublished = {\url{https://github.com/simomarsili/ndd}}, 
+	note = {[Accessed 22-10-2024]},
 } 
 
 @article{entropart,
@@ -195,5 +198,3 @@ doi = {10.1023/A:1026096204727}
  volume = {26},
  year = {2013}
 }
-
-
