+++ paper/paper.bib
@@ -148,32 +148,28 @@ doi = {10.1023/A:1026096204727}
   title = {PYM},
   author = {pillowlab},
   year = {},
-  howpublished = {\url{https://github.com/pillowlab/PYMentropy/?tab=readme-ov-file}}, 
-	note = {[Accessed 22-10-2024]},
+  howpublished = {\url{https://github.com/pillowlab/PYMentropy/?tab=readme-ov-file} [Accessed 2024-10-22]}, 
 } 
 
 @misc{bub,
 	author = {Liam Paninski},
 	title = {BUB},
-	howpublished = {\url{http://www.stat.columbia.edu/~liam/research/code/BUBfunc.m}},
+	howpublished = {\url{http://www.stat.columbia.edu/~liam/research/code/BUBfunc.m} [Accessed 2024-10-22]},
 	year = {},
-	note = {[Accessed 22-10-2024]},
 }
 
 @misc{nsb, 
   title = {NSB},
   author = {Ilya Nemenman},
   year = {},
-  howpublished = {\url{https://sourceforge.net/projects/nsb-entropy/}}, 
-	note = {[Accessed 22-10-2024]},
+  howpublished = {\url{https://sourceforge.net/projects/nsb-entropy/} [Accessed 2024-10-24]}, 
 } 
 
 @misc{ndd, 
   title = {NDD},
   author = {Simone Marsili},
   year = {},
-  howpublished = {\url{https://github.com/simomarsili/ndd}}, 
-	note = {[Accessed 22-10-2024]},
+  howpublished = {\url{https://github.com/simomarsili/ndd} [Accessed 2024-10-22]}, 
 } 
 
 @article{entropart,
