### 2024-10-23 kellino/DiscreteEntropy.jl#12 @kellino
@nluetts any better with the citations? I've included accessed date in the bib information, just in case the links ever die
