### 2024-10-23 kellino/DiscreteEntropy.jl#12 @nluetts
Yes, this has much improved!

For better looks, you can add spaces before the citations in lines 48, 65-73 and remove the surplus space after `[@chaoshen] , ...` in line 40. (Line numbers refer to markdown source file.)

Otherwise, why not cite the final paper from Hausser and Strimmer instead of the arxiv preprint? The Journal provides the bibtex code:

```
@article{JMLR:v10:hausser09a,
  author  = {Jean Hausser and Korbinian Strimmer},
  title   = {Entropy Inference and the {J}ames-{S}tein Estimator, with Application to Nonlinear Gene Association Networks},
  journal = {Journal of Machine Learning Research},
  year    = {2009},
  volume  = {10},
  number  = {50},
  pages   = {1469-1484},
  url     = {http://jmlr.org/papers/v10/hausser09a.html}
}
```

I also found that JOSS has an example on how to cite GitHub repos (https://joss.readthedocs.io/en/latest/example_paper.html). Accordingly, for `pym` and `ndd` you could do:

```
@misc{pym,
  author = {Pillowlab},
  title = {PYM entropy estimator MATLAB reference implementation},
  year = {2020},
  publisher = {GitHub},
  journal = {GitHub repository},
  url = {https://github.com/pillowlab/PYMentropy/},
}

@misc{ndd,
  author = {Simone Marsili},
  title = {{ndd} - Bayesian entropy estimation from discrete data},
  year = {2021},
  publisher = {GitHub},
  journal = {GitHub repository},
  url = {https://github.com/simomarsili/ndd},
}
```

For `nsb` you may want to use the title on sourceforge, ` title = {{NSB} {E}ntropy {E}stimation}`.

Finally, a colon after

```
These assumptions can have a profound effect on estimations of more complex measures, such as mutual information:
```

in line 101 would make it consistent with line 94.

That should be all with the cosmetics.
