### 2024-10-23 kellino/DiscreteEntropy.jl#12 @nluetts
So there is one more thing I noticed when reading through the paper a last time just now.

In lines 43 to 44 (in the paper, not source file) you write:

> The PYM(Pillowlab, 2020), BUB(Paninski) and Unseen(Valiant
> & Valiant, 2013) estimators are found ***only in the author’s Matlab implementions***.

I guess the PYM library is yours, but BUB was authored by Paninski and for Unseen a link to the matlab source is not provided (it does not seem to be included in PYM). Can you clarify?
