### 2024-10-23 7c146f4
fix link in CONTRIBUTING.md
