### 2024-10-23 8e9795b
do not assume user has github CLI (`gh`)
