### 2024-10-23 05aac95
make types in docs conform to types in code
