+++ src/InfoTheory/divergence.jl
@@ -70,9 +70,9 @@ Compute the [Kullback-Lebler Divergence](https://en.wikipedia.org/wiki/Kullback%
 between two discrete distributions. ``P`` and ``Q`` must be the same length.
 If the distributions are not normalised, they will be.
 
-If the distributions are not over the same space or the cross entropy is negative, then it returns ``Inf``.
+If the distributions are not over the same space or the cross entropy is negative, then it returns `Inf`.
 
-If truncate is set to some integer value, ```x```, return kl_divergence rounded to ```x``` decimal places.
+If truncate is set to some integer value, `x`, return kl_divergence rounded to `x` decimal places.
 """
 function kl_divergence(P::CountVector, Q::CountVector, estimator::Type{T}; truncate::Union{Nothing,Int}=nothing, α=0.0) where {T<:AbstractEstimator}
   if estimator == Bayes
