### 2024-10-23 kellino/DiscreteEntropy.jl#11 @nluetts
Hi @kellino

I ran your updated code against my script. For Unseen, the results are more realistic now, but still much worse than the other estimators and not what Contreras Rodríguez et al. report in their paper. It also fails some times:

```
[nluetts@fedora DiscreteEntropy-JOSS-Review]$ julia --project --load comparison.jl -e "main()"
###############################################################################

...

Running comparison for Unseen ...
Sample size: 8
Reference mean entropy: 5.202, DiscreteEntropy.jl: 0.06876407943033852
Sample size: 16
Reference mean entropy: 6.751, DiscreteEntropy.jl: 0.16074670482604692
Sample size: 32
Reference mean entropy: 7.739, DiscreteEntropy.jl: 0.3273587002393337
Sample size: 64
Reference mean entropy: 7.925, DiscreteEntropy.jl: 0.5499381518155609
Sample size: 128
Reference mean entropy: 7.945, DiscreteEntropy.jl: 0.9208798178505762
Sample size: 256
Reference mean entropy: 7.957, DiscreteEntropy.jl: 1.4577827444473084
Sample size: 512
Reference mean entropy: 7.95, DiscreteEntropy.jl: 2.162698415707029
Sample size: 1024
Reference mean entropy: 7.948, DiscreteEntropy.jl: 2.9027093645486657
Sample size: 2048
Error: unable to factorize the basis matrix (1)
ERROR running estimation: MathOptInterface.ResultIndexBoundsError{MathOptInterface.VariablePrimal}(MathOptInterface.VariablePrimal(1), 0)
Sample size: 4096
Error: unable to factorize the basis matrix (1)
ERROR running estimation: MathOptInterface.ResultIndexBoundsError{MathOptInterface.VariablePrimal}(MathOptInterface.VariablePrimal(1), 0)
Sample size: 8192
Error: unable to factorize the basis matrix (1)
ERROR running estimation: MathOptInterface.ResultIndexBoundsError{MathOptInterface.VariablePrimal}(MathOptInterface.VariablePrimal(1), 0)
Sample size: 16384
Reference mean entropy: 7.971, DiscreteEntropy.jl: 4.91103173101616
###############################################################################

┌───────────┬───────┬───────┬───────┬───────┬───────┬────────┬───────┬───────┬───────┬───────┬───────┬───────┐
│ Estimator │     8 │    16 │    32 │    64 │   128 │    256 │   512 │  1024 │  2048 │  4096 │  8192 │ 16384 │
├───────────┼───────┼───────┼───────┼───────┼───────┼────────┼───────┼───────┼───────┼───────┼───────┼───────┤
...
│    Unseen │ 0.069 │ 0.161 │ 0.327 │  0.55 │ 0.921 │  1.458 │ 2.163 │ 2.903 │   NaN │   NaN │   NaN │ 4.911 │
...
└───────────┴───────┴───────┴───────┴───────┴───────┴────────┴───────┴───────┴───────┴───────┴───────┴───────┘

```
