### 2024-10-24 kellino/DiscreteEntropy.jl#12 @kellino
Ah, an ambiguous sentence which I'll fix.  "_the  author's_..." does not refer to me (the author of the paper) but the code implementations by the creators of the original estimators. So the PYM estimator exists under the pillowlab user name on github. I'm unaware of any other implementation (except for the one included here). I should have a bit of free time tomorrow to make these last changes to the paper.
