### 2024-10-24 kellino/DiscreteEntropy.jl#11 @kellino
@nluetts, now, this is getting interesting. I looked at the Rodriguez paper again and they state for Unseen (on page 4) "the authors propose to compute its value algorithmically". This is a strange sentence and makes me wonder if they manually computed values (and perhaps got it wrong?). It's true that unseen source code doesn't seem to be available as a repo anywhere that I could find, but it is included as an appendix in the paper. I worked from this appendix.

As for the error when sample size gets too big... I'm not sure what the issue is there. Many of these estimators do come with quite strong assumptions, so perhaps Unseen has some upper limit. I'll investigate this further.
