### 2024-10-24 kellino/DiscreteEntropy.jl#11 @nluetts
Also with regard to the discrepancies for SG, SHR, JEF, LAP, MIN, BUB, maybe it is a good idea to contact the authors and ask them what could have happened here? Maybe it is just my testing implementation which is not exactly what they did, but then I wonder why the comparison succeeds for the remaining estimators.
