### 2024-10-24 kellino/DiscreteEntropy.jl#11 @kellino
I'll reach out, but I hope this submission won't be blocked waiting for their reply... there was an error in BUB which was recently fixed in a comment, this might reduce the discrepency there. The others  all seem to be those found in the R library, so perhaps there was a mistake in their R code?
