+++ paper/paper.bib
@@ -144,12 +144,14 @@ doi = {10.1023/A:1026096204727}
       primaryClass={stat.ML}
 }
 
-@misc{pym, 
-  title = {PYM},
-  author = {pillowlab},
-  year = {},
-  howpublished = {\url{https://github.com/pillowlab/PYMentropy/?tab=readme-ov-file} [Accessed 2024-10-22]}, 
-} 
+@misc{pym,
+  author = {Pillowlab},
+  title = {PYM entropy estimator MATLAB reference implementation},
+  year = {2020}, 
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  url = {https://github.com/pillowlab/PYMentropy/},
+}
 
 @misc{bub,
 	author = {Liam Paninski},
@@ -159,18 +161,21 @@ doi = {10.1023/A:1026096204727}
 }
 
 @misc{nsb, 
-  title = {NSB},
+  title = {{NSB} {E}ntropy {E}stimation},
   author = {Ilya Nemenman},
   year = {},
   howpublished = {\url{https://sourceforge.net/projects/nsb-entropy/} [Accessed 2024-10-24]}, 
 } 
 
 @misc{ndd, 
-  title = {NDD},
   author = {Simone Marsili},
-  year = {},
-  howpublished = {\url{https://github.com/simomarsili/ndd} [Accessed 2024-10-22]}, 
+  title = {{ndd} - Bayesian entropy estimation from discrete data},
+  year = {2021}, 
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  url = {https://github.com/simomarsili/ndd},
 } 
+
 
 @article{entropart,
   title = {{entropart}: An {R} Package to Measure and Partition Diversity},
