### 2024-10-25 kellino/DiscreteEntropy.jl#12 @kellino
@nluetts hopefully the ambiguous paragraph should be clearer now. Please let me know. I've also updated the bib file to the github references.
