### 2024-10-25 kellino/DiscreteEntropy.jl#12 @nluetts
@kellino yes, now the sentence is clear. I ticked off the remaining items in the "Software paper" part of my review list.
