### 2024-10-25 kellino/DiscreteEntropy.jl#11 @nluetts
Before I forget it (again), you have to get rid of these two lines:

https://github.com/kellino/DiscreteEntropy.jl/blob/44590e710a6827b692f3d8bb8cf8c5b326277558/src/Core/countdata.jl#L161

https://github.com/kellino/DiscreteEntropy.jl/blob/44590e710a6827b692f3d8bb8cf8c5b326277558/src/Core/countdata.jl#L163

When I run my script, the output is quite enormous ...
