### 2024-10-25 kellino/DiscreteEntropy.jl#11 @kellino
Ouch, I must have left those in for debugging. I'll remove now
