### 2024-10-25 593e6e7
fix: remove unnecessary print
