### 2024-10-25 kellino/DiscreteEntropy.jl#11 @kellino
Calls to print() removed in 6963a95cf9ba19ad8fc5e8405bef402e07dde25b
