### 2024-10-25 kellino/DiscreteEntropy.jl#11 @nluetts
`entropy` does not seem to include automatic tests, though. At least I as a non R-programmer are not able to spot them (https://github.com/cran/entropy/tree/master). But I assume they tested well enough manually.
