### 2024-10-25 kellino/DiscreteEntropy.jl#11 @nluetts
@kellino OK, so my last open question is `Unseen`.

You said there was a bug which was fixed. After the fix, the negative entropy values disappeared and the results are more reasonable. The problem is, that your testcase

https://github.com/kellino/DiscreteEntropy.jl/blob/6963a95cf9ba19ad8fc5e8405bef402e07dde25b/test/estimator_test.jl#L24

did not catch this bug. It succeeded already before the bugfix. So this test alone does not seem to be sufficient.

I think the paper from Valiant & Valiant can be used for another testcase, though perhaps not a quantitative one, since they only seem to report graphs, not numbers.

The first plot in their Figure 1 shows how the error of Unseen evolves with sample size for a discrete uniform distribution from 1 to 1000. In Julia, I can sample from this distribution like so:

```julia
julia> rand(1:1000, 100) # 100 samples
```

So to estimate the entropy:

```julia
julia> estimate_h(from_samples(svector(rand(1:1000, 100))), Estimator)
```

If I calculate $e$ to the power of the entropy, I should get the number of distinct items in the discrete distribution.
So if I oversample, the estimator should not have a problem to find that there are 1000 distinct numbers for this discrete distribution. And indeed:

```julia
julia> estimate_h(from_samples(svector(rand(1:1000, 10000))), PYM) |> exp
1007.6118832749003

julia> estimate_h(from_samples(svector(rand(1:1000, 10000))), NSB) |> exp
981.4373913617699

julia> estimate_h(from_samples(svector(rand(1:1000, 10000))), ChaoShen) |> exp
954.4725165955898
```

For less samples, the estimates deviate more, but are still close enough:

```julia
julia> using Statistics

julia> mean(estimate_h(from_samples(svector(rand(1:1000, 1000))), PYM) |> exp for _ in 1:500)
1381.6990603254876

julia> mean(estimate_h(from_samples(svector(rand(1:1000, 1000))), PYM) |> exp for _ in 1:500)
1379.908995113134

julia> mean(estimate_h(from_samples(svector(rand(1:1000, 1000))), NSB) |> exp for _ in 1:500)
629.4125878097738

julia> mean(estimate_h(from_samples(svector(rand(1:1000, 1000))), NSB) |> exp for _ in 1:500)
629.793778092218

julia> mean(estimate_h(from_samples(svector(rand(1:1000, 1000))), ChaoShen) |> exp for _ in 1:500)
1094.844219101877

julia> mean(estimate_h(from_samples(svector(rand(1:1000, 1000))), ChaoShen) |> exp for _ in 1:500)
1097.9925117939667
```

Now `Unseen`:

```julia
julia> estimate_h(from_samples(svector(rand(1:1000, 10000))), Unseen) |> exp
12.752808081795651

julia> estimate_h(from_samples(svector(rand(1:1000, 10000))), Unseen) |> exp
12.527995830000302

julia> estimate_h(from_samples(svector(rand(1:1000, 10000))), Unseen) |> exp
12.812724640864127

julia> mean(estimate_h(from_samples(svector(rand(1:1000, 1000))), Unseen) |> exp for _ in 1:500)
2.7727911829598364

julia> mean(estimate_h(from_samples(svector(rand(1:1000, 1000))), Unseen) |> exp for _ in 1:500)
2.7721745172018806
```

This result is very far from the ground truth and far from the claim in the paper that the error of the estimate should approach zero in case of this distribution and 1000 samples (upper left plot in Figure 1).

Can you check your code again and test against more examples with the original matlab code?
