### 2024-10-26 kellino/DiscreteEntropy.jl#11 @nluetts
> [...] It's true that unseen source code doesn't seem to be available as a repo anywhere that I could find, but it is included as an appendix in the paper. I worked from this appendix. [...]

In case you have not seen it:

https://theory.stanford.edu/~valiant/code.html

(Could also be added as a reference to the paper.)
