### 2024-10-28 ed1ce52
fix: more tests for unseen
