### 2024-10-28 kellino/DiscreteEntropy.jl#11 @kellino
@nluetts I've added more tests to catch the previously uncaught bug, but I'm not sure what's happening with the the most recent examples you've posted. It'll probably take me a few days to solve this, as time is a bit limited at the moment.
