+++ paper/paper.bib
@@ -167,6 +167,13 @@ doi = {10.1023/A:1026096204727}
   howpublished = {\url{https://sourceforge.net/projects/nsb-entropy/} [Accessed 2024-10-24]}, 
 } 
 
+@misc{unseenimp, 
+  title = {{U}nsee},
+  author = {Gregory Valiant},
+  year = {},
+  howpublished = {\url{https://theory.stanford.edu/~valiant/code.html} [Accessed 2024-10-24]}, 
+} 
+
 @misc{ndd, 
   author = {Simone Marsili},
   title = {{ndd} - Bayesian entropy estimation from discrete data},
@@ -199,3 +206,4 @@ doi = {10.1023/A:1026096204727}
  volume = {26},
  year = {2013}
 }
+
