+++ paper/paper.bib
@@ -168,7 +168,7 @@ doi = {10.1023/A:1026096204727}
 } 
 
 @misc{unseenimp, 
-  title = {{U}nsee},
+  title = {{U}nseen},
   author = {Gregory Valiant},
   year = {},
   howpublished = {\url{https://theory.stanford.edu/~valiant/code.html} [Accessed 2024-10-24]}, 
