### 2024-11-01 eee45d4
fix: stupid bug in unseen
