+++ paper/paper.bib
@@ -169,7 +169,7 @@ doi = {10.1023/A:1026096204727}
 
 @misc{unseenimp, 
   title = {{U}nseen},
-  author = {Gregory Valiant},
+  author = {Valiant, Paul and Valiant, Gregory},
   year = {},
   howpublished = {\url{https://theory.stanford.edu/~valiant/code.html} [Accessed 2024-10-24]}, 
 } 
