### 2024-11-01 edaecc3
updated tests for unseen
