### 2024-11-01 kellino/DiscreteEntropy.jl#11 @kellino
@nluetts a genuinely stupid bug that took me a while to actually find, but only revealed itself when the sample size / support size ration became relatively large. This produces a much saner answer now.

```
julia> mean(estimate_h(from_samples(svector(rand(1:1000, 10000))), Unseen) |> exp for _ in 1:500)
990.5554343074119
```
