### 2024-11-01 02b7a05
more tests on divergence
