### 2024-11-01 kellino/DiscreteEntropy.jl#11 @nluetts
@kellino
Great! And now it also still works quite well for less samples, as advertised in the paper:

```julia
julia> mean(estimate_h(from_samples(svector(rand(1:1000, 1000))), Unseen) |> exp for _ in 1:500)
987.5433915762957

julia> mean(estimate_h(from_samples(svector(rand(1:1000, 1000))), Unseen) |> exp for _ in 1:500)
987.8677858039398

julia> mean(estimate_h(from_samples(svector(rand(1:1000, 100))), Unseen) |> exp for _ in 1:500)
1074.078845471124

julia> mean(estimate_h(from_samples(svector(rand(1:1000, 100))), Unseen) |> exp for _ in 1:500)
1027.9313289893134
```
