### 2024-11-01 #7334 @nluetts
@kellino fixed all issues that came up during my review, so I finished my checklist.

The package performs as described in the paper, it comes with comprehensive documentation, and should be a very helpful tool for people working with entropy estimation. @Nikoleta-v3 I recommend the paper for acceptance.
