### 2024-11-05 #7334 @kellino
@Nikoleta-v3 I just wanted to check what else is needed from me at this point, or should I just wait for a decision?
