### 2024-11-11 #7334 @Nikoleta-v3
Well, this (https://github.com/openjournals/joss-reviews/issues/7334#issuecomment-2458929644) didn’t age well 😅 @kellino, I’ve opened a pull request and I just saw that you accepted the changes 👍🏻

In Markdown, you can set the code environment to Julia for better syntax highlighting 😄

At this point could you please:

- [x] (if necessary) Make a new tagged release of your software, and list the version tag of the archived version here. Or let me know if I should just use the release  v0.2.0
