### 2024-11-11 #7334 @kellino
@Nikoleta-v3 I think a small version bump to v0.2.1 is in order as there were many bug fixes. I've uploaded the software to
https://doi.org/10.5281/zenodo.14067392 with v0.2.1 in the Project.toml metadata. Hopefully everything else should be correct!
