### 2024-11-11 #7334 @kellino
> @kellino, on Zenodo (https://doi.org/10.5281/zenodo.14067392), you used the "Creative Commons Attribution" license, but your software on GitHub is under the MIT license. I believe JOSS requires them to be the same 😄

Done!
