### 2024-11-12 #7334 @Nikoleta-v3
```
- 10.1007/s10664-024-10556-3 may be a valid DOI for title: Hyperfuzzing: black-box security hypertesting with...
- 10.2307/j.ctv2rr3j2g.3 may be a valid DOI for title: BUB
```

I have checked and the above suggestions are incorrect 😄
