### 2024-11-12 #7334 @kellino
@Nikoleta-v3 the doi for Hyperfuzzing is correct (and new as of last week) so please give me a bit of time to put that into the paper draft.
