### 2024-11-12 #7334 @Nikoleta-v3
Thank you @kellino! For some reason when I search for the DOI this publication came up: https://dl.acm.org/doi/10.1145/3551349.3556947 but I can see now that it is the correct one
