### 2024-11-12 7d3d89e
fix missing doi and duplicate cite
