+++ paper/paper.bib
@@ -23,11 +23,16 @@ eprint = {https://doi.org/10.1137/1104033},
  doi = {10.1145/3368089.3409748}
 }
 
-@article{blackwell2023hyperfuzzing,
+@article{blackwell2025hyperfuzzing,
   title={Hyperfuzzing: black-box security hypertesting with a grey-box fuzzer},
   author={Blackwell, Daniel and Becker, Ingolf and Clark, David},
-  journal={arXiv preprint arXiv:2308.09081},
-  year={2023}
+  journal={Empirical Software Engineering},
+  volume={30},
+  number={1},
+  pages={1--28},
+  year={2025},
+  doi={10.1007/s10664-024-10556-3},
+  publisher={Springer}
 }
 
 @book{MacKay2003,
@@ -105,15 +110,6 @@ doi = {10.1023/A:1026096204727}
     doi = {10.1162/NECO_a_00266},
     url = {https://doi.org/10.1162/NECO\_a\_00266},
     eprint = {https://direct.mit.edu/neco/article-pdf/24/5/1368/865674/neco\_a\_00266.pdf},
-}
-
-@misc{hausser2009entropy,
-      title={Entropy inference and the {J}ames-{S}tein estimator, with application to nonlinear gene association networks},
-      author={Jean Hausser and Korbinian Strimmer},
-      year={2009},
-      eprint={0811.3579},
-      archivePrefix={arXiv},
-      primaryClass={stat.ML}
 }
 
 @book{Cover2006,
