### 2024-11-12 #7334 @kellino
@Nikoleta-v3 both of those things should now be fixed.
