### 2024-11-12 #7334 @danielskatz
@kellino - I've suggested a number of small changes in https://github.com/kellino/DiscreteEntropy.jl/pull/17. Please merge this, or let me know what you disagree with.

Also, I note that there is no Acknowledgments section in your paper. Should there be (to acknowledge funding or other support)?
