### 2024-11-12 #7334 @kellino
The only financial support was my salary, but that should probably be mentioned. I'll put that in now
