### 2024-11-12 #7334 @danielskatz
Congratulations to @kellino (David Kelly) and co-author on your publication!!

JOSS depends on volunteers and we couldn't do this work without you
